% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_anchors.R
\name{anchors_APA}
\alias{anchors_APA}
\title{Anchor positions for APA}
\usage{
anchors_APA(ABS, RES, bedpe, dist_thres = c(0, Inf), mode = c("both",
  "cis", "trans"))
}
\arguments{
\item{ABS}{The indices slot of a GENOVA experiment.}

\item{RES}{The resolution slot of a GENOVA experiment.}

\item{bedpe}{A \code{data.frame} with 6 columns in BEDPE format containing the
regions to be anchored: chrom1/start1/end1/chrom2/start2/end2.}

\item{dist_thres}{An \code{integer} vector of length 2 indicating the minimum
and maximum distances in basepairs between anchorpoints.}

\item{mode}{A \code{character} vector of length 1 indicating which
interaction to take. Possible values: \code{"cis"}, \code{"trans"} or
\code{"both"}.}
}
\value{
A \code{matrix} with two columns.
}
\description{
Transforms a BEDPE formatted \code{data.frame} to indices of the Hi-C matrix.
}
\details{
The resulting matrix contains ordered indices to the Hi-C matrix
  slot in the GENOVA experiment.

  The \code{mode} argument determines what entries are
  reported. \code{"cis"} returns entries within chromosomes;
  \code{"trans"} gives these between chromosomes and \code{"bot"}.
}
\seealso{
\code{\link[GENOVA]{APA}} for context.

  \code{\link[GENOVA]{bed2idx}} for general conversion of BED-like
  \code{data.frame}s to Hi-C indices.
}
