% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_anchors.R
\name{anchors_ATA}
\alias{anchors_ATA}
\title{Anchor positions for ATA}
\usage{
anchors_ATA(ABS, bed, dist_thres = c(225000, Inf), padding = 1)
}
\arguments{
\item{ABS}{The indices slot of a GENOVA experiment.}

\item{bed}{A \code{data.frame} with 3 columns in BED format, containing TAD
boundary positions for one TAD at each row.}

\item{dist_thres}{An \code{integer} vector of length 2 indicating the minimum
and maximum TAD sizes to include.}

\item{padding}{A \code{numeric} of length 1 to determine the padding around
TADs, expressed in TAD widths.}
}
\value{
A \code{matrix} with two columns.
}
\description{
Transforms a BED-formatted \code{data.frame} containing TAD positions into
indices to the Hi-C matrix.
}
\details{
The resulting matrix contains ordered indices to the Hi-C matrix
  slot in the GENOVA experiment.
}
\seealso{
\code{\link[GENOVA]{ATA}} for context.

  \code{\link[GENOVA]{bed2idx}} for general conversion of BED-like
  \code{data.frame}s to Hi-C indices.
}
