% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_anchors.R
\name{anchors_filter_oob}
\alias{anchors_filter_oob}
\title{Filter anchors that become out of bounds.}
\usage{
anchors_filter_oob(ABS, anchors, rel_pos)
}
\arguments{
\item{ABS}{The indices slot of a GENOVA experiment.}

\item{anchors}{A \code{matrix} with two columns containing pre-computed
anchor indices.}

\item{rel_pos}{An \code{integer} vector indicating relative positions in
bins, indicating ranges around anchors to lookup.}
}
\value{
A \code{matrix} with two columns.
}
\description{
Discards anchors for which a lookup will result in any lookup bins that are
out of bounds.
}
\details{
The resulting matrix contains indices to the Hi-C matrix slot in the
  GENOVA experiment.

  An index is considered out of bounds when that index plus the maximum, or
  minus the minimum, relative position would belong to a different chromosome.
}
