% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_anchors.R
\name{anchors_finish}
\alias{anchors_finish}
\title{Finish anchors for repeated matrix lookup}
\usage{
anchors_finish(IDX, anchors, rel_pos, shift = 0)
}
\arguments{
\item{IDX}{The indices slot of a GENOVA \code{contacts} object.}

\item{anchors}{A \code{matrix} with two columns containing pre-computed
anchor indices. See the \code{\link[GENOVA]{anchors}} documentation.}

\item{rel_pos}{An \code{integer} vector indicating relative positions in
bins, indicating ranges around anchors to lookup.}

\item{shift}{An \code{integer} of length 1 indicating how many bins the
anchors should be shifted.}
}
\value{
A \code{anchors} object of the same type.
}
\description{
Finish anchors for repeated matrix lookup
}
\examples{
\dontrun{
anch <- anchors_APA(WT_20kb$ABS, WT_20kb$RES, loops)
anchors_finish(WT_20kb$ABS, anch, -10:10, 0)
}
}
