% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_anchors.R
\name{anchors_PESCAn}
\alias{anchors_PESCAn}
\title{Anchor positions for PE-SCAn}
\usage{
anchors_PESCAn(ABS, RES, bed, dist_thres = c(5000000L, Inf),
  min_compare = 10L, mode = c("cis", "trans", "both"))
}
\arguments{
\item{ABS}{The indices slot of a GENOVA experiment.}

\item{RES}{The resolution slot of a GENOVA experiment.}

\item{bed}{A \code{data.frame} with 3 columns in BED format, containing the
regions to anchor in pairwise manner.}

\item{dist_thres}{An \code{integer} vector of length 2 indicating the minimum
and maximum distances in basepairs between anchorpoints.}

\item{min_compare}{An \code{integer} of length 1 indicating the minimum
number of pairwise interactions on a chromosome to consider.}

\item{mode}{A \code{character} vector of length 1 indicating which
interaction to take. Possible values: \code{"cis"}, \code{"trans"} or
\code{"both"}.}
}
\value{
A \code{matrix} with two columns.
}
\description{
Takes all pairwise interactions of locations on the genome and maps these to
indices of the Hi-C matrix.
}
\details{
The resulting matrix contains ordered indices to the Hi-C matrix
  slot in the GENOVA experiment.

  The \code{mode} argument determines what pairwise interactions are
  reported. \code{"cis"} returns pairwise interactions within a chromosome;
  \code{"trans"} gives these between chromosomes and \code{"bot"}.
}
\seealso{
\code{\link[GENOVA]{PESCAn}} for context.

  \code{\link[GENOVA]{bed2idx}} for general conversion of BED-like
  \code{data.frame}s to Hi-C indices.
}
