% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateBEDPEwithBED.R
\name{annotateBEDPEwithBED}
\alias{annotateBEDPEwithBED}
\title{Annotate loops}
\usage{
annotateBEDPEwithBED(bedpe, bed, annotCol, gap = 0, verbose = F)
}
\arguments{
\item{bedpe}{Loop-dataframe with at least six columns}

\item{bed}{Bed-file, containing a column with annotations}

\item{annotCol}{Which column-number should we take for the annotation?}

\item{gap}{Maximal allowed gap between bedpe-bed overlaps. Default = 0.}

\item{verbose}{Should this function be chatty?}
}
\value{
An annotated bedpe-dataframe
}
\description{
Using some genomicRanges-magic, annotate bedpe-files with a bed file.
}
