% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{antoni}
\alias{antoni}
\title{Antoni}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 350 rows and 350 columns.
}
\usage{
antoni
}
\description{
The face of an angel, the curls of the sea.
}
\keyword{datasets}
