% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA.R
\name{APA}
\alias{APA}
\title{Get a z-stack matrix of 2D-regions (e.g. loops).}
\usage{
APA(experiment, loop.bed, smallTreshold = 225000, verbose = F)
}
\arguments{
\item{experiment}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{loop.bed}{Bedpe file containing the loop positions: produced by read.bedpe().}

\item{smallTreshold}{The minimal size of loops. Too small loops lead to messy plots.}

\item{verbose}{Produces a progress-indication.}
}
\value{
A tidy data.frame containing the Z-stack scores.
}
\description{
Extract matrices around a defined set of pixels, like possible loops.
Sums over all matrices to produce a single Z-stack matrix, which is normalised to 100 loops.
}

