% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_visualise.R
\name{RMT_visualisation}
\alias{RMT_visualisation}
\alias{visualise.APA_discovery}
\alias{APA_visualisation}
\alias{PESCAn_visualisation}
\alias{visualise.PESCAn_discovery}
\title{Visualise APA and PE-SCAn results}
\usage{
\method{visualise}{APA_discovery}(results, subtract = 1, raw = FALSE)

\method{visualise}{PESCAn_discovery}(results, subtract = 1,
  mode = c("obsexp", "signal"), raw = FALSE)
}
\arguments{
\item{results}{A results object as returned by the \code{APA} function or the
\code{PESCAn} function.}

\item{subtract}{An \code{integer} or \code{character} matching an experiment
name of length 1, specifying which sample should be subtracted to make the
difference panels. Alternatively, set to \code{NULL} for no difference
panels.}

\item{raw}{A \code{logical} of length 1: should a bare mimimum plot be
returned?}

\item{mode}{(PE-SCAn only) A \code{character} vector of length 1 with
\code{"obsexp"} for plotting the observed over expected or \code{"signal"}
for plotting the signal only (no background correction).}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot APA and PE-SCAn results, by default with comparisons of differences.
}
\details{
Difference (\u0394) panels are created by subtracting the values of each
  sample by the values of the sample indicated by the \code{subtract}
  argument. Hence, a clear positive signal in the difference panels indicates
  enrichment in that sample versus the \code{subtract}-sample.

  The \code{raw = TRUE} argument allows customisation of the plot: it returns
  a plot object with no position or fill scales and no theme settings.

  When \code{raw = TRUE} and \code{subtract} is not \code{NULL}, the plot
  will not render unless a continuous fill scale is supplied with the
  argument \code{aesthetics = "fill2"}. This a small trade-off for having two
  independent fill scales in a single plot.
}
\examples{
# APA
apa <- APA(list(WT = WT_40kb, KO = KO_40kb), WT_loops)
autoplot(apa)

# PE-SCAn
pescan <- PESCAn(list(WT = WT_40kb, KO = KO_40kb), super_enhancers)
autoplot(pescan)

# To plot PE-SCAn without background correction
autoplot(pescan, mode = "signal")

# Handling 'raw' plots
autoplot(pescan, raw = TRUE) +
  ggplot2::scale_fill_gradient(aesthetics = "fill2")
}
\seealso{
\code{\link[GENOVA]{APA}} and \code{\link[GENOVA]{PESCAn}}
}
