% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMLA.R
\name{ARA}
\alias{ARA}
\title{Aggregate Region Analysis}
\usage{
ARA(
  explist,
  bed,
  shift = 1e+06,
  size_bin = NULL,
  size_bp = NULL,
  outlier_filter = c(0, 1),
  anchors = NULL,
  raw = TRUE
)
}
\arguments{
\item{explist}{Either a single GENOVA \code{contacts} object or a list of
GENOVA \code{contacts} objects.}

\item{bed}{A BED-formatted \code{data.frame} with the following 3 columns:
\enumerate{ \item A \code{character} giving the chromosome names. \item An
\code{integer} with start positions. \item An \code{integer} with end
positions. } Note that rows wherein the second column is larger than the
third column are considered to be in the reverse direction.}

\item{shift}{An \code{integer} of length 1 indicating how many basepairs the
anchors should be shifted. Essentially performs circular permutation of
\code{size} for a reasonable estimate of background. The argument is
ignored when \code{shift <= 0}.}

\item{size_bin, size_bp}{The size of the lookup regions in bins (i.e. a score
of 21 yields an output with 10 Hi-C bins both up- and downstream of the
anchor). When \code{NULL} (default), it is internally set to \code{21} when
the \code{size_bp} is also \code{NULL}. \code{size_bp} is an alternative
parametrisation for the lookup regions, expressed in basepairs.
\code{size_bp} is not used when the argument \code{size_bin} is set.}

\item{outlier_filter}{A \code{numeric} of length 2 between \code{[0-1]}
indicating quantiles of data to be used as thresholds. Data outside these
thresholds are set to the nearest threshold. Setting this to \code{c(0, 1)}
performs no outlier correction.}

\item{anchors}{(Optional) A \code{matrix} with two columns containing
pre-computed anchor indices. If this is set, skips calculation of anchor
indices and uses this argument instead. See
\code{\link[GENOVA:anchors]{anchors_ARA()}}.}

\item{raw}{A \code{logical} of length 1: should the raw array underlying the
summary matrices be returned in the output? Should be \code{TRUE} if the
intention is to use the \code{quantify} function.}
}
\value{
An \code{ARA_discovery} object containing the following slots:
  \describe{ \item{obsexp}{An \code{array} with the dimensions
  \code{size_bin} x \code{size_bin} x \code{length(explist)} containing fold
  change values for the signal over the (banded) median, shifted values.}
  \item{signal}{An \code{array} with the dimensions \code{size_bin} x
  \code{size_bin} x \code{length(explist)} containing mean contact values for
  bins surrounding the anchors.} \item{signal_raw}{A \code{list} with
  \code{length(explist)} elements for each contacts object, wherein an
  element is an n x \code{size_bin} x \code{size_bin} array with contact
  values for each anchor. 'n' is the number of non-empty valid anchors.}
  \item{shifted}{An \code{array} with the dimensions \code{size_bin} x
  \code{size_bin} x \code{length(explist)} containing mean contact values for
  bins that are \code{shift} basepairs away from the anchors.}
  \item{shifted_raw}{A \code{list} with \code{length(explist)} elements for
  each contacts object, wherein an alemeent is an n x \code{size_bin} x
  \code{size_bin} array with contact values for each shifted anchors. 'n' is 
  the number of non-empty valid (unshifted) anchors.}}
}
\description{
Extracts Hi-C matrices centered around regions and averages the results for
all regions. Can take orientations of regions into account.
}
\details{
By default, \code{ARA} also calculates the results for shifted
  anchors and normalises the \code{"obsexp"} slot by off-diagonal bands.

  The '\code{bed}' argument can take in oriented entries, wherein entries
  with a start site larger than the end site are considered to be in the
  reverse direction. The reverse sites are flipped during analysis, so the
  orientation is the same as in the forward sites.
}
\section{Resolution recommendation}{
 10kb-40kb
}

\examples{
\dontrun{
# Typical usage
ara <- ARA(list(WT_20kb, KO_20kb), ctcf_sites)

# Alternative usage with pre-calculated anchors
anchors <- anchors_ARA(WT_20kb$IDX, ctcf_sites)
ara <- ARA(list(WT_20kb, KO_20kb))

# Visualisation
visualise(ara)
}
}
\seealso{
The \code{\link[GENOVA]{rep_mat_lookup}} function that performs the
  lookup and summary for the \code{ARA} function and others. \cr The
  \code{\link[GENOVA]{discovery}} class for a general description of
  \code{discovery} classes. \cr The \code{\link[GENOVA]{visualise}} function
  for visualisation of the results. \cr The \code{\link[GENOVA]{quantify}}
  function for quantification of ARA results. \cr The
  \code{\link[GENOVA]{anchors}} documentation for more information about
  anchors.

Other aggregate repeated matrix lookup analyses: 
\code{\link{APA}()},
\code{\link{ATA}()},
\code{\link{CSCAn}()},
\code{\link{PESCAn}()},
\code{\link{rep_mat_lookup}()}
}
\concept{aggregate repeated matrix lookup analyses}
