% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_anchors.R
\name{as_anchors}
\alias{as_anchors}
\title{Coerce to anchors}
\usage{
as_anchors(x)
}
\arguments{
\item{x}{A two column \code{matrix} or \code{data.frame} with integers.}
}
\value{
An object of class \code{anchors}.
}
\description{
Function to coerce to \code{anchors} if possible.
}
\details{
The resulting \code{anchors} has the default '\code{type}' attribute
  \code{"custom"}.
}
\examples{
\dontrun{
as_anchors(matrix(1:20, 2))
}
}
\seealso{
\code{\link[GENOVA]{anchors}}
}
