% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badBin.find.R
\name{badBin.find}
\alias{badBin.find}
\title{Find all bad bins}
\usage{
badBin.find(experiment, p.treshold = 0.1, plotBad = F, getBed = T)
}
\arguments{
\item{experiment}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{p.treshold}{Cutoff of spearmans correlation.}

\item{getBed}{Produce a bed-like object.}
}
\value{
A list containing a vector with the bad bin-id's (`badBins`). Optionally, it can output a bed-file (`bed`).
}
\description{
Builds a representative RCP and tests per Hi-C bin, whether its log-linear fitted data correlates.
}

