% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bundeling.R
\name{bundle}
\alias{bundle}
\alias{bundle.ARMLA_discovery}
\alias{bundle.domainogram_discovery}
\alias{bundle.IS_discovery}
\alias{bundle.virtual4C_discovery}
\alias{bundle.CS_discovery}
\alias{bundle.saddle_discovery}
\alias{bundle.RCP_discovery}
\alias{bundle.DI_discovery}
\alias{bundle.IIT_discovery}
\alias{bundle.chrommat_discovery}
\title{Combine discovery objects}
\usage{
bundle(..., collapse = "_")

\method{bundle}{ARMLA_discovery}(..., collapse = "_")

\method{bundle}{domainogram_discovery}(..., collapse = "_")

\method{bundle}{IS_discovery}(..., collapse = "_")

\method{bundle}{virtual4C_discovery}(..., collapse = "_")

\method{bundle}{CS_discovery}(..., collapse = "_")

\method{bundle}{saddle_discovery}(..., collapse = "_")

\method{bundle}{RCP_discovery}(..., collapse = "_")

\method{bundle}{DI_discovery}(..., collapse = "_")

\method{bundle}{IIT_discovery}(..., collapse = "_")

\method{bundle}{chrommat_discovery}(..., collapse = "_")
}
\arguments{
\item{...}{A number of \code{discovery} objects from the same analysis type.}

\item{collapse}{An optional character string to seperate names in case the
experiment names are not unique. Not \code{NA_character_}.}
}
\value{
A \code{discovery}-class object of the same type.
}
\description{
\code{bundle} takes one or more GENOVA \code{discovery} objects
  and combines them along experiments.
}
\details{
This function can be convenient when a the same analysis has been
  run on a sequence of experiments with parameters specific to the
  experiment. For example, consider the case when \code{ATAs} have been run
  on several single experiments with TAD boundary anchors that match that
  same experiment. The \code{bundle} function allows combining the resulting
  \code{discovery} objects for visualisation or quantification.

  The arrays in slots of the \code{discovery} objects need to have compatible
  first and second dimensions, i.e. they must have had the same
  '\code{size_bp}' or '\code{size_bin}' arguments and have been run at the
  same resolutions.
}
\examples{
\dontrun{
# Running multiple analysis
ata1 <- ATA(WT_10kb, tads_wt)
ata2 <- ATA(KO_10kb, tads_ko)

# Combining results
cata <- bundle(WT = ata1, KO = ata2)

# Visualising the combined results
visualise(cata)
}
}
\seealso{
The \code{\link[GENOVA]{discovery}} class.
  \code{\link[GENOVA]{unbundle}} for splitting \code{discovery} objects.
}
