% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centromere_telomere.R
\name{centromere.telomere.analysis}
\alias{centromere.telomere.analysis}
\title{Calculate a scaled matrix of the interchromosomal interactions aligned from centromere to telomere}
\usage{
centromere.telomere.analysis(exp, chrom.vec, nrow = 100, leave.out = NULL,
  q.top = 1e-05, verbose = F)
}
\arguments{
\item{exp}{GENOVA data structure for your experiment of interest}

\item{chrom.vec}{a vector containing the chromosomes that should be considered in the pair-wise interchromosomal interactions}

\item{nrow}{dimensions of the matrix (number of columns will be the same)}

\item{leave.out}{two-column matrix or data.frame containing the specific chromosome combinations that need to left out}

\item{q.top}{top quantile of scores that should be left out of the analysis (because they are outliers)}

\item{verbose}{Produces a progress-indication.}
}
\description{
Calculate a scaled matrix of the interchromosomal interactions aligned from centromere to telomere
}
