% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_compat_exp}
\alias{check_compat_exp}
\title{Check compatability of a list of GENOVA experiments}
\usage{
check_compat_exp(explist)
}
\arguments{
\item{explist}{Either a single GENOVA \code{contacts} object or a list of
GENOVA \code{contacts} objects.}
}
\value{
A \code{list} of GENOVA experiment(s).
}
\description{
Checks if the indices (ABS slot) across experiments are identical.
}
\keyword{internal}
