% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.chrom.comparison.R
\name{chrom.comparison.plot}
\alias{chrom.comparison.plot}
\title{plot.chrom.comparison}
\usage{
chrom.comparison.plot(mat, color.fun = NULL, z.max = NULL)
}
\arguments{
\item{mat}{Output of chromosomeMatrix}

\item{color.fun}{Optional color-function}

\item{z.max}{Adjust the maximum value of the color-scale}
}
\value{
A plot of the matrix
}
\description{
Plot a genome-wide overview-matrix.
}
