% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_plot.R
\name{cis.compartment.plot}
\alias{cis.compartment.plot}
\title{Draw intrachromosomal interaction heatmap for a chromosome (arm) with corresponding compartment scores}
\usage{
cis.compartment.plot(exp, chrom, arm = "p", cut.off = NULL, obs.exp = F,
  invert = F, color.scheme = "fall", cs.lim = NULL)
}
\arguments{
\item{exp}{GENOVA data object}

\item{chrom}{chromosome that should be drawn}

\item{arm}{which chromosome arm: 'p' or 'q' (for acrocentric this should be set to 'q')}

\item{cut.off}{maximum value for the heatmap}

\item{obs.exp}{whether an observed over expected matrix should be drawn (default is false)}

\item{invert}{whether the compartment score should be inverted}

\item{color.scheme}{color scheme that should be used, defaults to fall, other values result in white-red gradient}

\item{cs.lim}{y-axis limit for the compart score, if unset (NULL) will default to the maximum absolute value}
}
\description{
Draw intrachromosomal interaction heatmap for a chromosome (arm) with corresponding compartment scores
}
