% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cis_trans.R
\name{cis_trans}
\alias{cis_trans}
\title{Get the percentage cis-contacts genome-wide or from a region}
\usage{
cis_trans(exp, bed = NULL)
}
\arguments{
\item{exp}{A contacts-objects or a list of them}

\item{bed}{A bed-dataframe or a (named) list of them.}
}
\value{
A data.table with the sample, percentage and region.
}
\description{
Get the percentage cis-contacts genome-wide or from a region
}
\section{Resolution recommendation}{
 500kb-1Mb
}

