% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cisTotal.perChrom.R
\name{cisTotal.perChrom}
\alias{cisTotal.perChrom}
\title{cisTotal.perChrom}
\usage{
cisTotal.perChrom(exp, ...)
}
\arguments{
\item{exp}{The Hi-C experiment object of a sample: produced by construct.experiment().}
}
\value{
The per-chromosome cis/total ratio
}
\description{
Calculate the ratio of the intrachromosomal interactions over the interchromosomal interactions
}
\note{
The per chromsosome score is on average lower than the whole genome score, because in the per chromosome analysis every ditag is counted twice (for the two interacting chromosomes)
}
