% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cisTotal.scores.R
\name{cisTotal.scores}
\alias{cisTotal.scores}
\title{cisTotal.scores}
\usage{
cisTotal.scores(exp)
}
\arguments{
\item{exp}{The Hi-C experiment object of a sample: produced by construct.experiment().}
}
\value{
The genome-wide cis/total ratio
}
\description{
Calculate the ratio of the intrachromosomal interactions over the interchromosomal interactions
}

