% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.experiment.R
\name{construct.experiment}
\alias{construct.experiment}
\title{Construct a HiC-experiment.}
\usage{
construct.experiment(ICEDpath, BEDpath, SAMPLENAME, COLOR = 1,
  COMMENTS = NULL)
}
\arguments{
\item{ICEDpath}{Full path to HiC-pro matrix-file.}

\item{BEDpath}{Full path the HiC-pro index-file}

\item{SAMPLENAME}{The name of the sample.}

\item{COLOR}{Color associated with sample.}

\item{COMMENTS}{A place to store some comments.}
}
\value{
A list.
}
\description{
Make a structure which holds the most needed information of a HiC-experiment.
}

