% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov2d.R
\name{cov2d}
\alias{cov2d}
\title{cov2d}
\usage{
cov2d(experiment, bed, minDist = 5e+06, maxDist = Inf, size = 5e+05,
  add = 0, outlierCutOff = 40, rmOutlier = F, verbose = T)
}
\arguments{
\item{experiment}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{bed}{A bed-dataframe.}

\item{minDist}{The minimal distance}

\item{size}{The amount of Hi-C bins to take into account (i.e. a score of 21 yield an output with 10 Hi-C bins up- and downstream of the anchor).}

\item{add}{Shift pairs with a constant value to bed-start and -end.}

\item{outlierCutOff}{The severity of outliers: roughly translates to the amount of MADs above the median.}

\item{rmOutlier}{Try to perform outlier-correction}

\item{verbose}{Verbose?}

\item{minDist}{The maximal distance}
}
\value{
A list containing a score-matrix and a count-variable of the amount of used bed-entries.
}
\description{
Get scores around given positions
}
