% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMLA.R
\name{CSCAn}
\alias{CSCAn}
\title{Cross spatial chromatin analysis}
\usage{
CSCAn(
  explist,
  bedlist,
  shift = 1000000L,
  dist_thres = c(NA, 1e+06),
  size_bin = NULL,
  size_bp = NULL,
  outlier_filter = c(0, 1),
  min_compare = 10,
  anchors = NULL,
  raw = FALSE
)
}
\arguments{
\item{explist}{Either a single GENOVA \code{contacts} object or a list of
GENOVA \code{contacts} objects.}

\item{bedlist}{A \code{list} of length >= 2 wherein each element is a
BED-like \code{data.frame} containing three columns for chromosome, start-
and end-positions.}

\item{shift}{An \code{integer} of length 1 indicating how many basepairs the
anchors should be shifted. Essentially performs circular permutation of
\code{size} for a reasonable estimate of background. The argument is
ignored when \code{shift <= 0}.}

\item{dist_thres}{An \code{integer} vector of length 2 indicating the minimum
and maximum distances in basepairs between anchorpoints.}

\item{size_bin}{The size of the lookup regions in bins (i.e. a score of 21
yields an output with 10 Hi-C bins both up- and downstream of the anchor).
When \code{NULL} (default), it is internally set to \code{21} when the
\code{size_bp} is also \code{NULL}.}

\item{size_bp}{Alternative parametrisation for the lookup regions, expressed
in basepairs. Not used when the argument \code{size_bin} is set.}

\item{outlier_filter}{A \code{numeric} of length 2 between \code{[0-1]}
indicating quantiles of data to be used as thresholds. Data outside these
thresholds are set to the nearest threshold. Setting this to \code{c(0, 1)}
performs no outlier correction.}

\item{min_compare}{An \code{integer} of length 1 indicating the minimum
number of pairwise interactions on a chromosome to consider.}

\item{anchors}{(Optional) A \code{matrix} with two columns containing
pre-computed anchor indices. If this is set, skips calculation of anchor
indices and uses this argument instead.}

\item{raw}{A \code{logical} of length 1: should the raw array underlying the
summary matrices be returned in the output? Should be \code{TRUE} if the
intention is to use the \code{quantify} function.}
}
\value{
A \code{CSCAn_discovery} object containing the results of the C-SCAn.
}
\description{
Takes a list of BED-like genomic locations and makes a Hi-C contact analysis
for crosswise pairs between elements of that list.
}
\examples{
\dontrun{
NULL
}
}
\seealso{
The \code{\link[GENOVA]{rep_mat_lookup}} function that performs the
  lookup and summary for the \code{CSCan} function and others. \cr The
  \code{\link[GENOVA]{discovery}} class for a general description of
  \code{discovery} classes. \cr The \code{\link[GENOVA]{visualise}} function
  for visualisation of the results. \cr The
  \code{\link[GENOVA]{anchors}} documentation for more information about
}
