% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.TAD.scatterplot.R
\name{differential.TAD.scatterplot}
\alias{differential.TAD.scatterplot}
\title{differential.TAD.scatterplot}
\usage{
differential.TAD.scatterplot(exp1, exp2, color.fun = NULL, xlab = NULL,
  ylab = NULL, ...)
}
\arguments{
\item{exp1}{Output of intra.inter.TAD.contacts for experiment 1.}

\item{exp2}{Output of intra.inter.TAD.contacts for experiment 2.}

\item{color.fun}{A color-function like rainbow().}

\item{xlab}{The x-axis label.}

\item{ylab}{The y-axis label.}
}
\value{
A scatterplot.
}
\description{
Create a scatter plot of the inter and intra TAD contact frequencies.
}
