% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_discovery.R
\name{discovery}
\alias{discovery}
\title{Discovery class}
\description{
The discovery class contains the results of the analysis
  functions within GENOVA.
}
\details{
Next to the results, the \code{discovery} objects may also contain
  useful metadata on the analysis that was run, such as which resolution was
  used.

  Functions that generate \code{discovery} objects are the following:
  \describe{
   \item{\code{\link[GENOVA]{PESCAn}}}{\code{PESCAn_discovery} objects}
    \item{\code{\link[GENOVA]{APA}}}{\code{APA_discovery} objects}
    \item{\code{\link[GENOVA]{ATA}}}{\code{ATA_discovery} objects}
    \item{\code{\link[GENOVA]{ARA}}}{\code{ARA_discovery} objects}
  }
}
\section{Operations}{
 \subsection{Subsetting}{\code{discovery} objects can be
  subsetted by using \code{subset(discovery, i)} wherein \code{i} is an
  \code{integer} or \code{character} corresponding to the intended
  sample(s).} \subsection{Combining}{\code{discovery} objects of the same
  type can be combined by using \code{\link[GENOVA]{bundle}(discovery_A,
  discovery_B)}.} \subsection{Splitting}{\code{discovery} objects can be
  split to individual samples by using
  \code{\link[GENOVA]{unbundle}(discovery)}.}
  \subsection{Visualisation}{\code{discovery} objects can be visualised using
  \code{\link[GENOVA]{visualise}(discovery)}}
  \subsection{Quantification}{\code{discovery} objects can be quantified
  using \code{\link[GENOVA]{quantify}(discovery)}}
}

