% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centromere_telomere.R
\name{draw.centromere.telomere}
\alias{draw.centromere.telomere}
\title{draw the log2-ratio of the relative interaction frequencies
the ratio is calculated over the median of the matrix}
\usage{
draw.centromere.telomere(m, cut.off = 2)
}
\arguments{
\item{m}{matrix result from centromere-telomere}

\item{cut.off}{log2-ratio cut.off}
}
\description{
draw the log2-ratio of the relative interaction frequencies
the ratio is calculated over the median of the matrix
}
