% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_matrix}
\alias{dt_matrix}
\title{Upper triangle sparse symmetric triplet matrix to dense matrix}
\usage{
dt_matrix(x, i, j, dim, offset)
}
\arguments{
\item{x}{A \code{numeric} vector containing the data for matrix elements.}

\item{i}{A \code{integer} vector parallel to \code{x} containing row
positions.}

\item{j}{A \code{integer} vector parallel to \code{x} containing column
positions.}

\item{dim}{A \code{integer} of length 1 specifying the dimensions of the
output matrix.}

\item{offset}{A \code{integer} of length 1 noting a potential offset in the
  \code{i} and \code{j} arguments.

  Particularly useful when lookup up square regions around the diagonal in
  Hi-C data.}
}
\value{
A \code{dim * dim} sized \code{matrix}.
}
\description{
Convenience function for returning square, dense symmetric matrices from the
upper triangle of triplet format symmetric matrices.
}
\note{
For reasons of speed, no checks are performed wether the input is
  compatible with sensible output.
}
\keyword{internal}
