% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{erase_GENOVA_cache}
\alias{erase_GENOVA_cache}
\title{Erase cache.}
\usage{
erase_GENOVA_cache()
}
\value{
\code{0} for success and \code{1} for failure.
}
\description{
GENOVA can download test data with the
\code{\link[GENOVA]{get_test_data}()} function. This function undoes that
by deleting the files in the local cache. Test data can then be re-downloaded
using the \code{get_test_data(..., download = TRUE)} function.
}
\examples{
NULL
}
