% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_expnames.R
\name{expnames}
\alias{expnames}
\alias{expnames<-}
\title{Sample names for an object}
\usage{
expnames(x, simplify = TRUE)

expnames(x) <- value
}
\arguments{
\item{x}{The object for which to retrieve sample names.}

\item{simplify}{A \code{logical} of length 1: if the \code{x} argument is a
list, should the sample names be returned as a flat character?}

\item{value}{A non-\code{NA} \code{character} vector of appropriate length.}
}
\value{
A \code{character} or \code{list} with sample names
}
\description{
Looks in the appropriate spot for sample names and returns or sets these.
}
\examples{
\dontrun{
expnames(discovery_object)
expnames(contacts_object)
}
}
