% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastDI.R
\name{fastDI}
\alias{fastDI}
\title{Get DI-scores for a specific region}
\usage{
fastDI(experiment, chrom, start, end, max.i = 100)
}
\arguments{
\item{experiment}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{chrom}{Chromosome.}

\item{start}{Start position in bp.}

\item{end}{End position in bp.}

\item{max.i}{Maximal size of flanking regions to calculate DI.}
}
\value{
A vector of DI-scores per Hi-C bin.
}
\description{
Extracts a sub-matrix from given coordinates and calculates the Directionality-Index.
}
