% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.interp.surface.R
\name{fields.interp.surface}
\alias{fields.interp.surface}
\title{fields.interp.surface}
\usage{
fields.interp.surface(obj, loc)
}
\arguments{
\item{obj}{A surface or image  object like the list for contour, persp or image.}

\item{lof}{A matrix of 2 d locations -- new points to evaluate the surface.}
}
\value{
An interpolated matrix
}
\description{
fields.interp.surface
}

