% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centromeres.R
\name{find_centromeres}
\alias{find_centromeres}
\title{Empirical centromeres}
\usage{
find_centromeres(MAT, IDX)
}
\arguments{
\item{MAT}{The \code{MAT} slot of a GENOVA \code{contacts} object.}

\item{IDX}{The \code{IDX} slot of a GENOVA \code{contacts} object.}
}
\value{
A \code{data.table} with chromosome, start-index and end-index
  columns.
}
\description{
For each chromosome, finds and reports the largest stretch of unused bins as
centromeres.
}
\details{
A bin is seen as unused when the index has an entry in the
  \code{IDX} argument but does not have entries in the \code{MAT} argument.
}
\examples{
\dontrun{
centros <- find_centromeres(exp$MAT, exp$IDX)
}
}
