% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.from.matrix.R
\name{frequency.from.matrix}
\alias{frequency.from.matrix}
\title{Plot pyramid-style HiC-regions.}
\usage{
\method{frequency}{from.matrix}(experiment, chrom, start, end,
  window = 10000, q.val = 0.95, cut.off = NULL,
  extraPadding = 1.75, ylim = c(0, 250), shinyAxis = T, ...)
}
\arguments{
\item{experiment}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{chrom}{Chromosome.}

\item{start}{Start position in bp.}

\item{end}{End position in bp.}

\item{window}{The window-size (default: 10e3).}

\item{q.val}{Quantile cutoff}

\item{cut.off}{Score cutoff}

\item{extraPadding}{Extra room needed for the ends of the piramid.}

\item{ylim}{The y-start and -stop: zero is at the diagonal,}
}
\value{
A plot.
}
\description{
Extracts a sub-matrix from given coordinates and plots them pyramid-style (i.e. the Hi-C diagonal becomes horizontal).
}
