% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation_score.R
\name{genome.wide.insulation}
\alias{genome.wide.insulation}
\title{genome.wide.insulation}
\usage{
genome.wide.insulation(hic, window.size, normalize.genome = F, verbose = F)
}
\arguments{
\item{hic}{exp}

\item{window.size}{The sliding square size}

\item{verbose}{Chatty?}
}
\value{
DF with insulation score
}
\description{
Get genome-wide insulation
}
