% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GENOVA.R
\docType{package}
\name{GENOVA-package}
\alias{GENOVA}
\alias{GENOVA-package}
\title{GENOVA: GENOVA: expore the Hi-C}
\description{
GENOVA consist of a set of analysis- and visualisation-functions. These will help to discover valuable information buried deep in your HiC-matrices.
}
\author{
\strong{Maintainer}: Robin H. van der Weide \email{r.vd.weide@nki.nl}

Authors:
\itemize{
  \item Teun van den Brand \email{t.vd.brand@nki.nl}
  \item Elzo de Wit \email{e.d.wit@nki.nl}
}

}
\keyword{internal}
