% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{get_test_data}
\alias{get_test_data}
\title{Test data}
\usage{
get_test_data(res, download = FALSE)
}
\arguments{
\item{res}{The resolution of test data to use. Either \code{"40k"} or
\code{"150k"}.}

\item{download}{A \code{logical(1)}. If \code{FALSE} (default), will not
download the data if it cannot be found in the cache. If \code{TRUE}, data
is downloaded when not found in the cache.}
}
\value{
A GENOVA \code{contacts} at 40k or 150k resolution.
}
\description{
Retrieves a test \code{contacts} object from a local cache at either 40k or
150k resolution. The test data only contains human chromosomes 21 and 22.
}
\details{
The data originates from wildtype Hap1 cells from the Haarhuis
  \emph{et al}. (2017) publication. The full data is available in raw fastq
  files or valid pair files at GEO accession GSE95014. The data were mapped
  against the 'hg19' reference genome using HiC-Pro v2.7.7.

  The test data only needs to be downloaded only once as it is stored in a
  local cache. The local cache can be cleared using the
  \code{\link[GENOVA]{erase_GENOVA_cache}()} function.
}
\examples{
\dontrun{
# Fails the first time
exp <- get_test_data("40k")

# Should succees if `download = TRUE`
exp <- get_test_data("40k", download = TRUE)

# From now on, it should succeed
exp <- get_test_data("40k")

# Additional calls with `download = TRUE` do not re-download the data.
exp <- get_test_data("40k", download = TRUE)
}
}
