% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCseg.callTAD.R
\name{HiCseg.callTAD}
\alias{HiCseg.callTAD}
\title{HiCseg.callTAD}
\usage{
HiCseg.callTAD(
  experiment,
  chromsToUse = NULL,
  binPerBorder = 3,
  chunk = NULL,
  BEDcolor = "255,255,0",
  verbose = F
)
}
\arguments{
\item{experiment}{The Hi-C experiment object: produced by construct.experiment().}

\item{chromsToUse}{The chromosome(s) of interest. Omitting this leads to the computation across all chromosomes.}

\item{binPerBorder}{Minimum number of bins per border: used to set the maximal
number of change-points. The default (3) set this maximal number of borders to nbins/3.}

\item{chunk}{Use chunks of [chunk]bp instead of the whole arm: the whole arm can be a big burden,
so it could be beneficial to use chunks of a set size (e.g. 5e6). But be carefull:
the segmentation will be different from the full arm!!!}

\item{BEDcolor}{Color of items in the resulting BEDPE-file}

\item{verbose}{Bioinformatics can be a bit lonely: set this to true to get a more chatty function.}
}
\value{
A BEDPE-df
}
\description{
A wrapper with call TADs with HiCseg of Lévy-Leduc et al. (2018)
}
