% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation.callTAD.plot.R
\name{insulation.callTAD.plot}
\alias{insulation.callTAD.plot}
\title{insulation.callTAD.plot}
\usage{
insulation.callTAD.plot(exp, calls, chrom, start, stop, cut.off)
}
\arguments{
\item{calls}{Output from insulation.callTAD}

\item{chrom}{Chromosome of interest}

\item{start}{Start of interest}

\item{stop}{Stop of interest}

\item{cut.off}{Zlim of matrix}

\item{exp1}{The Hi-C experiment object: produced by construct.experiment().}
}
\value{
A plot
}
\description{
Plot called TADs
}

