% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation_functions.R
\name{insulation.domainogram_old}
\alias{insulation.domainogram_old}
\title{insulation.domainogram}
\usage{
insulation.domainogram_old(
  exp,
  chrom,
  start,
  end,
  axes = T,
  window.size1 = 5,
  window.size2 = 101,
  step = 2,
  local = T,
  zlim = c(-1, 0.5)
)
}
\arguments{
\item{exp}{A Hi-C experiment object: produced by loadContacts().}

\item{chrom}{Chromosome of the region of interest}

\item{start}{Start position of the region of interest}

\item{end}{End position of the region of interest}

\item{axes}{Plot axes}

\item{window.size1}{The minimal sliding square size}

\item{window.size2}{The maximal sliding square size}

\item{step}{Thesliding square step-size}

\item{local}{Local or per-chromosome normalisation?}

\item{zlim}{Zlims of the plot}
}
\value{
A plot
}
\description{
create a domainogram for the insulation scores
}
\examples{
# Make an domainogram of a locus on chromosome 7 with windowsizes from 1 to 101.
\dontrun{
insulation.domainogram(Hap1_WT_10kb, 
                       chrom = 'chr7', 
                       start = 25.5e6, 
                       end = 30e6, 
                       window.size1 = 1, 
                       window.size2 = 101, 
                       step = 2)
}
}
