% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation.heatmap.R
\name{insulation.heatmap}
\alias{insulation.heatmap}
\title{Plot insulation-heatmaps.}
\usage{
insulation.heatmap(insulationList, bed, focus = 1, sortWidth = 10,
  whatToPlot = "both", profileFunct = mean, flank = 5e+05, zlim = c(-1,
  1), profileCols = NULL, profileZlim = NULL, heatmapCols = c("#f03b20",
  "#ffeda0", "white", "#31a354"), rmNonVariateRows = T, verbose = F)
}
\arguments{
\item{insulationList}{A named list of results from `genome.wide.insulation()`.}

\item{bed}{A bed-df with TAD-calls. Only the first three columns are important.}

\item{focus}{Sort on which sample? (give the index-number)}

\item{sortWidth}{Percentage of columns to sort on (aligned on center)}

\item{whatToPlot}{Do you want to plot a profile, heatmap or both?}

\item{profileFunct}{Function to make profile-plots (mean, median, sum, absSum)}

\item{flank}{Amount of flanking bp.}

\item{zlim}{zlims c(min,max)}

\item{profileCols}{Vector of colors for profiles}

\item{profileZlim}{zlims c(min,max) for profiles}

\item{heatmapCols}{a vector of colors for the heatmap-gradient.}

\item{rmNonVariateRows}{ignore rows with an sd of 0}

\item{verbose}{Chatty?}
}
\value{
A plot
}
\description{
insulations in, plot out
}
