% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation_score.R
\name{insulation.plot.dual}
\alias{insulation.plot.dual}
\title{plot.insulation.dual}
\usage{
insulation.plot.dual(exp1, exp2, chrom, start, end, cut.off = NULL,
  window.size = 21, local = T, delta = F)
}
\arguments{
\item{exp1}{A Hi-C experiment object: produced by construct.experiment().}

\item{exp2}{A Hi-C experiment object: produced by construct.experiment().}

\item{chrom}{Chromosome}

\item{start}{Start position of the region of interest}

\item{end}{End position of the region of interest}

\item{cut.off}{The cut.off for Hi-C scores}

\item{window.size}{The sliding square size}

\item{local}{Local or per-chromosome normalisation?}

\item{delta}{Plot the differential insulation scores (exp1 - exp2)}
}
\value{
A plot
}
\description{
Plot insulation-scores of two experiments, complete with Hi-C matrices, for 
a region of interest.
}
\examples{
# Make a matrix-plot, with two experiments for a region of interest.
\dontrun{
insulation.plot.dual(exp1 = Hap1_WT_10kb, 
                     exp2 = Hap1_WAPL_10kb, 
                     chrom = 'chr7', 
                     start = 25.5e6, 
                     end = 30e6, 
                     window.size = 21)
}
}
