% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation_score.R
\name{insulation.polygon}
\alias{insulation.polygon}
\title{insulation.domainogram}
\usage{
insulation.polygon(ins.score, rotate = F)
}
\arguments{
\item{ins.score}{Scores.}
}
\value{
happiness
}
\description{
create a domainogram for the insulation scores
}

