% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation_domainogram.R
\name{insulation_domainogram}
\alias{insulation_domainogram}
\title{Domainogram of insulation}
\usage{
insulation_domainogram(
  explist,
  chrom,
  start,
  end,
  window_range = c(5, 101),
  step = 2
)
}
\arguments{
\item{explist}{Either a single GENOVA \code{contacts} object or list of
GENOVA \code{contacts} objects.}

\item{chrom}{One of the following: \itemize{
  \item{A \code{character} of length one indicating a chromosome name.}
  \item{A 3-column, 1-row \code{data.frame} in BED-format.}
  \item{A single \code{character} describing a locus in UCSC notation, e.g. 
  \code{"chr1:30,000,000-40,000,000"}.}
} The latter two options automatically provide the \code{start} and 
\code{end} arguments too.}

\item{start, end}{An \code{integer} of length 1 with a region's start and end 
position in basepairs.}

\item{window_range}{An \code{integer} vector of length 2 noting the minimum
and maximum size of the sliding square.}

\item{step}{An \code{integer} of length 1 with the step size for incrementing
the size of the sliding square.}
}
\value{
A \code{domainogram_discovery} object containing a \code{data.frame}
  with the insulation scores for the region of interest at various sizes of
  the sliding square.
}
\description{
Creates a domainogram of insulation scores for a genomic region of interest
by calculating the insulation scores for a range of sliding square sizes.
}
\examples{
\dontrun{
# Making a domainogram
domgram <- insulation_domainogram(list(WT = WT_20kb, KO = KO_20kb),
                                  chrom = "chr7", start = 25e6, end = 30e6)
                                 
# Plotting the domainogram
visualise(domgram)
}
}
