% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixplot_insulation.R
\name{insulation_matrixplot}
\alias{insulation_matrixplot}
\title{Insulation matrix plot.}
\usage{
insulation_matrixplot(
  exp1,
  exp2 = NULL,
  IS_discovery = NULL,
  chrom,
  start,
  end,
  colour_lim = NULL,
  rasterise = FALSE,
  colour_bar = FALSE
)
}
\arguments{
\item{exp1}{A GENOVA \code{contacts} object.}

\item{exp2}{A GENOVA \code{contacts} object compatible with \code{exp1} 
(optional) or \code{NULL}. If not \code{NULL}, the \code{exp2} argument 
will be displayed alongside the \code{exp1} data.}

\item{IS_discovery}{A \code{IS_discovery} object, as created by the 
\code{\link[GENOVA]{insulation_score}} function. When \code{NULL}, the 
insulation score will be calculated on the fly for this region only.}

\item{chrom}{One of the following: \itemize{
  \item{A \code{character} of length 1 giving the chromosome name of the 
  region of interest.}
  \item{A 3-column, 1-row \code{data.frame} in BED-format.}
  \item{A single \code{character} describing a locus in UCSC notation, e.g.
  \code{"chr1:30,000,000-40,000,000"}.}
} The latter two options automatically provide the \code{start} and 
\code{end} arguments too.}

\item{start, end}{An \code{integer} with the start- and end-positions of the
region of interest in basepairs.}

\item{colour_lim}{A \code{numeric} of length two giving the minimum and 
maximum values in contact values. Out of bounds values will be set to the
nearest limits.}

\item{rasterise}{Set to \code{TRUE} to use a bitmap raster instead of 
polygons.}

\item{colour_bar}{A \code{logical} of length 1, indicating whether a
colour-bar legend should be drawn at the right.}
}
\value{
No values are returned but a plot is outputted on the graphics 
  device.
}
\description{
Display a Hi-C matrix with insulation scores annotated in the margins.
}
\details{
The 'on-the-fly' calculation of the insulation score uses the 
 default settings. Therefore, if for example the window size is to be 
 changed, please use the \code{insulation_score()} function and provide the
 result as the \code{IS_discovery} argument in this function.

 Because the 'on-the-fly' calculation of the insulation score only
 computes the insulation score for this locus, there might a small offset
 to the scores due to the limited scope for normalisation.
}
\section{Recommended resolution}{
 10kb-40kb
}

\examples{
\dontrun{
insulation_matrixplot(
  WT_20kb, KO_20kb,
  chrom ="chr1:40,000,000-60,000,000"
)
}
}
\seealso{
The \code{\link[GENOVA]{insulation_score}} function for computation
  of insulation scores.

Other matrix plots: 
\code{\link{compartment_matrixplot}()},
\code{\link{hic_matrixplot}()},
\code{\link{pyramid}()},
\code{\link{trans_matrixplot}()}
}
\concept{matrix plots}
