% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation_functions.R
\name{insulation_plot_single}
\alias{insulation_plot_single}
\title{plot_insulation_single}
\usage{
insulation_plot_single(
  exp,
  chrom,
  start,
  end,
  IS_discovery = NULL,
  colour_fun = NULL,
  cut.off = NULL,
  window.size = 21
)
}
\arguments{
\item{exp}{The Hi-C experiment object: produced by construct.experiment().}

\item{chrom}{Chromosome}

\item{start}{Start position of the region of interest}

\item{end}{End position of the region of interest}

\item{IS_discovery}{(Optional) Provide pre-calculated insulation scores.}

\item{colour_fun}{A function that generates a colour vector of length
\code{n} when given an integer. Typically the result of a call to
\code{colorRampPalette()}.}

\item{cut.off}{The cut.off for Hi-C scores}

\item{window.size}{The sliding square size}
}
\value{
A plot
}
\description{
Plot a single insulation-score and Hi-C matrix for a region of interest.
}
\examples{
# Make a matrix-plot of an experiment and the insulation-score for a region of interest.
\dontrun{
insulation.plot.single(exp1 = WT_10kb,
                       chrom = 'chr7',
                       start = 25.5e6,
                       end = 30e6,
                       window.size = 21)
}
}
