% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation_score.R
\name{insulation_score}
\alias{insulation_score}
\title{Calculate insulation scores}
\usage{
insulation_score(explist, window = 30, norm_to = c("chromosome",
  "genome"))
}
\arguments{
\item{explist}{Either a single GENOVA \code{contacts} object or list of
GENOVA \code{contact} objects.}

\item{window}{An \code{integer} of length 1 for the size of the sliding
square.}

\item{norm_to}{A \code{character} of length 1, either \code{"chromosome"} or
\code{"genome"} noting whether normalisation should occur per chromosome or
for the genome as a whole.}
}
\value{
A \code{IS_discovery} object
}
\description{
Insulation scores are calculated by sliding a square along the diagonal of
the Hi-C matrix and averaging the values in that square.
}
\details{
Typically, when the sliding square passes a TAD it yields high
  insulation scores whereas in between TADs it yields low insulation scores.
  Hence, TAD boundaries can be identified as local minima in the insulation
  score.

  The normalisation is performed by taking the
  \ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}} value over median,
  wherein the median is either calculated per chromosome or across the
  genome.
}
\examples{
\dontrun{
iscore <- insulation_score(explist(WT_20kb, KO_20kb), window = 20) 
}
}
