% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra.inter.TAD.contacts.R
\name{intra.inter.TAD.contacts}
\alias{intra.inter.TAD.contacts}
\title{intra.inter.TAD.contacts}
\usage{
intra.inter.TAD.contacts(exp, TAD, max.neighbor = 10)
}
\arguments{
\item{exp}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{TAD}{A bed-dataframe.}

\item{max.neighbor}{How many surrounding TADs should be taken into account?}
}
\value{
A list containing a dataframe with scores between two TADs and an updated version of the TAD-input file, which now also includes TAD-IDs.
}
\description{
Combine a bed file with TAD positions and the Hi-C matrix file to calculate the average coverage over a TAD and between two TADs.
}
