% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_anchors.R
\name{is_anchors}
\alias{is_anchors}
\title{Test object is anchors}
\usage{
is_anchors(x)
}
\arguments{
\item{x}{Any \R object.}
}
\value{
A \code{logical} vector of length 1.
}
\description{
Function to check if an object is \code{anchors}.
}
\examples{
\dontrun{
m <- matrix(1:20, 10)
is_anchors(m) # FALSE

m <- as_anchors(m)
is_anchors(m) # TRUE
}
}
\seealso{
\code{\link[GENOVA]{anchors}}
}
