% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_contacts.R
\name{load_contacts}
\alias{load_contacts}
\title{Construct a contacts-object.}
\usage{
load_contacts(
  signal_path,
  indices_path = NULL,
  resolution = 10000,
  sample_name = NULL,
  centromeres = NULL,
  colour = NULL,
  z_norm = FALSE,
  scale_bp = 1e+09,
  scale_cis = FALSE,
  balancing = TRUE,
  legacy = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{signal_path}{Full path to an Hi-C file. This could be \*.matrix (for HiCpro), \*.cooler or \*.hic (for juicer).}

\item{indices_path}{Full path the HiC-pro-like index-file. Required when `signal_path` is \*.matrix.}

\item{resolution}{Set the desired resolution of the matrix when using juicer-data.}

\item{sample_name}{The name of the sample.}

\item{centromeres}{One of the following: \itemize{
 \item \code{NULL} (default) for empirically estimating centromeres by 
 searching for the largest stretch of empty bins.
 \item A \code{data.frame} with three columns per chromosome:
 chromosome name, start-position and end-positions of centromeric regions.
 \item \code{FALSE} to set no centromere information.
}}

\item{colour}{colour associated with sample.}

\item{z_norm}{Normalise the matrices per-chromosome with a Z-score.}

\item{scale_bp}{Scale contacts to have genome-wide sum of `scale_bp` reads (default: 1e9). Set to NULL to skip this.}

\item{scale_cis}{Only scale with cis-contacts.}

\item{balancing}{TRUE (default) will perform matrix balancing for .cooler and KR for.hic.}

\item{legacy}{Get a pre-v1 object (mimics the output of construct.experiment.)}

\item{verbose}{Do you want updates during the construction?}
}
\value{
An contacts-object, which is a named list of contacts, indices and 
attributes for a Hi-C matrix of a given sample at a given resolution.
}
\description{
Make an object which holds the most needed information of a HiC-experiment at a given resolution.
}
\note{
Some reference genomes have very small "random" or "patch" chromosomes,
which can have zero contacts mapped to it (at certain resolutions).
`load_contacts` checks this and omits these chromosomes in the resulting
experiment-object. The RMCHROM-flag will also be set to TRUE: this will
help other GENOVA-functions to deal better with this problem. There is a
slight performance-cost during the construction of the object, however.
}
\examples{
\dontrun{
WT_10kb_hicpro <- load_contacts(signal_path = "WT_10kb_iced.matrix", 
                                indices_path = "WT_10kb_abs.bed", 
                                sample_name = "WT", 
                                colour = "black")
WT_10kb_cooler <- load_contacts("WT_10kb.cooler", 
                                balancing =T, 
                                sample_name = "WT", 
                                colour = "black")
WT_10kb_juicer <- load_contacts("WT_10kb_iced.hic", 
                                resolution = 10e3, 
                                balancing =T, 
                                sample_name = "WT", 
                                colour = "black")
}
}
