% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadContacts.R
\name{loadContacts}
\alias{loadContacts}
\title{Construct a contacts-object.}
\usage{
loadContacts(signalPath, indicesPath = NULL, juicerResolution = 10000,
  sampleName, centromeres = NULL, colour = NULL, Znorm = F,
  bpscaling = 1e+09, cisscaling = F, balancing = T, legacy = F,
  verbose = T)
}
\arguments{
\item{signalPath}{Full path to an Hi-C file. This could be \*.matrix (for HiCpro), \*.cooler or \*.hic (for juicer).}

\item{indicesPath}{Full path the HiC-pro-like index-file. Required when `signalPath` is \*.matrix.}

\item{juicerResolution}{Set the desired resolution of the matrix when using juicer-data.}

\item{sampleName}{The name of the sample.}

\item{centromeres}{A data.frame with three columns per chromosome: chromosome name, start-position and end-position of the centromeric region.}

\item{colour}{colour associated with sample.}

\item{Znorm}{Normalise the matrices per-chromosome with a Z-score.}

\item{balancing}{TRUE (default) will perform matrix balancing for .cooler and KR for.hic.}

\item{legacy}{Get a pre-v1 object (mimics the output of construct.experiment.)}

\item{verbose}{Do you want updates during the construction?}

\item{BPscaling}{Scale contacts to have agenome-wide sum of `BPscaling` reads (default: 1e9). Set to NULL to skip this.}
}
\value{
An contacts-object, which is a named list of contacts, indices and attributes for a Hi-C matrix of a given sample at a given resolution.
}
\description{
Make an object which holds the most needed information of a HiC-experiment at a given resolution.
}
\note{
Some reference genomes have very small "random" or "patch" chromosomes,
which can have zero contacts mapped to it (at certain resolutions).
`loadContacts` checks this and omits these chromosomes in the resulting
experiment-object. The RMCHROM-flag will also be set to TRUE: this will
help other GENOVA-functions to deal better with this problem. There is a
slight performance-cost during the construction of the object, however.
}
\examples{
WT_10kb_hicpro <- loadContacts(signalPath = "WT_10kb_iced.matrix", indicesPath = "WT_10kb_abs.bed", sampleName = "WT", colour = "black")
WT_10kb_cooler <- loadContacts("WT_10kb.cooler", balancing =T, sampleName = "WT", colour = "black")
WT_10kb_juicer <- loadContacts("WT_10kb_iced.hic", juicerResolution = 10e3, balancing =T, sampleName = "WT", colour = "black")
}
