% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation_score.R
\name{log2overmean}
\alias{log2overmean}
\title{Log 2 value over its mean}
\usage{
log2overmean(x)
}
\arguments{
\item{x}{A \code{numeric} vector}
}
\value{
A \code{numeric} vector with the calculated values.
}
\description{
A small convenience function to calculate the log 2 of a value divided by the
mean of this set of values.
}
\details{
Ignores \code{NA} values in calculation of the mean
}
\examples{
log2overmean(runif(100))
}
