% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation_score.R
\name{log2overmedian}
\alias{log2overmedian}
\title{Log 2 value over its median}
\usage{
log2overmedian(x)
}
\arguments{
\item{x}{A \code{numeric} vector}
}
\value{
A \code{numeric} vector with the calculated values.
}
\description{
A small convenience function to calculate the log 2 of a value divided by the
median of this set of values.
}
\details{
Ignores \code{NA} values in calculation of the median.
}
\examples{
log2overmedian(runif(100))
}
