% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation_score.R
\name{matrix.insulation}
\alias{matrix.insulation}
\title{chromosome.wide.insulation}
\usage{
matrix.insulation(mat, window.size)
}
\arguments{
\item{mat}{exp}

\item{window.size}{The sliding square size}
}
\value{
happiness
}
\description{
get insulation for matrix
}

