% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_matrix_lookup.R
\name{matrix_lookup}
\alias{matrix_lookup}
\title{Engine for repeated matrix lookups}
\usage{
matrix_lookup(ICE, anchors, rel_pos)
}
\arguments{
\item{ICE}{The Hi-C matrix slot of a GENOVA experiment object.}

\item{anchors}{A \code{matrix} with two columns containing pre-computed
anchor indices.}

\item{rel_pos}{An \code{integer} vector indicating relative positions in
bins, indicating ranges around anchors to lookup.}
}
\value{
A threedimensional \code{array} wherein the first dimension is
  parallel to the rows in \code{anchors}, and the second and third dimensions
  are of length \code{m}.
}
\description{
It looks up parts of the Hi-C matrix of a single experiment.
}
\seealso{
\code{\link[GENOVA]{rep_mat_lookup}}
}
\keyword{internal}
