% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMLA.R
\name{parse_rel_pos}
\alias{parse_rel_pos}
\title{Parse size and resolution to relative positions}
\usage{
parse_rel_pos(res, size_bin, size_bp)
}
\arguments{
\item{res}{The RES slot of a GENOVA Hi-C experiment.}

\item{size_bin}{The size of the lookup regions in bins (i.e. a score of 21
yields an output with 10 Hi-C bins both up- and downstream of the anchor).}

\item{size_bp}{Alternative parametrisation for the lookup regions, expressed
in basepairs. Not used when the argument \code{size_bin} is set.}
}
\value{
A \code{integer} vector of relative positions.
}
\description{
Parse size and resolution to relative positions
}
\keyword{internal}
