% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PESCAn.R
\name{PESCAn}
\alias{PESCAn}
\title{PE-SCAn}
\usage{
PESCAn(exp, bed, shift = 1e+06, mindist = 5e+06, size = 4e+05,
  rmOutlier = F)
}
\arguments{
\item{exp}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{bed}{A bed-dataframe.}

\item{shift}{Set to X bp for circular permutation. Set to zero for just getting the signal-matrix.}

\item{size}{Size in bp of window.}

\item{rmOutlier}{Try to perform outlier-correction}

\item{minDist}{The minimal distance}
}
\value{
A O/E score-matrix.
}
\description{
From a ChIP bed file, a HiCpro matrix and a HiCpro bed file calculate a PE-scan like data structure. Run this per chromosome.
}
\examples{
# Run PE-SCAn on a bed of super-enhancers, using WT Hi-C data.
SE <- read.delim(superEnhancers.bed, header = F)

WT_PE = PESCAn_run(exp = WT_exp, bed = SE)
MUT_PE = PESCAn_run(exp = MUT_exp, bed = SE)

list_PE = list("WT" = WT_PE, 'Mutant' = MUT_PE)
PESCAn_plot(list_PE, 4e4, title = 'PE-SCAn', zTop = c(1,1.5))

}
