% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PESCAn_APA.R
\name{PESCAn2}
\alias{PESCAn2}
\title{Paired-end spatial chromatin analysis}
\usage{
PESCAn2(explist, bed, shift = 1000000L, dist_thres = c(5000000L, Inf),
  size_bin = NULL, size_bp = 4e+05, outlier_filter = c(0, 1),
  min_compare = 10, anchors = NULL, raw = FALSE)
}
\arguments{
\item{explist}{Either a single GENOVA experiment or a list of GENOVA
experiments.}

\item{bed}{A \code{data.frame} with 3 columns in BED format, containing the
regions to anchor in pairwise manner.}

\item{shift}{An \code{integer} of length 1 indicating how many basepairs the
anchors should be shifted. Essentially performs circular permutation of
\code{size} for a reasonable estimate of background. The argument is
ignored when \code{shift <= 0}.}

\item{dist_thres}{An \code{integer} vector of length 2 indicating the minimum
and maximum distances in basepairs between anchorpoints.}

\item{size_bin}{The size of the lookup regions in bins (i.e. a score of 21
yields an output with 10 Hi-C bins both up- and downstream of the anchor).}

\item{size_bp}{Alternative parametrisation for the lookup regions, expressed
in basepairs. Not used when the argument \code{size_bin} is set.}

\item{outlier_filter}{A \code{numeric} of length 2 between \code{[0-1]}
indicating quantiles of data to be used as thresholds. Data outside these
thresholds are set to the nearest threshold. Setting this to \code{c(0, 1)}
performs no outlier correction.}

\item{min_compare}{An \code{integer} of length 1 indicating the minimum
number of pairwise interactions on a chromosome to consider.}

\item{anchors}{(Optional) A \code{matrix} with two columns containing
pre-computed anchor indices. If this is set, skips calculation of anchor
indices and uses these instead.}

\item{raw}{A \code{logical} of length 1: should the raw array underlying
the summary matrices be returned in the output?}
}
\value{
A \code{list} of the same length as \code{explist} wherein list
  elements contain the results of the PE-SCAn per experiment.
}
\description{
Performs an all-to-all Hi-C contact analysis for specified regions.
}
\examples{
# Typical usage: PESCAn for super enhancers using a 1 MB
# circular permutation on a pair of experiments.
pescan <- PESCAn2(explist = list(WT_40kb, KO_40kb),
                  bed = super_enhancers,
                  shift = 1e6)

# Alternative usage with pre-calculated anchors and no permutation
anchors <- anchors_PESCAn(WT_40kb$ABS, WT_40kb$RES,
                          genes_tss,
                          dist_thres = c(5e6, 15e6))
pescan <- PESCAn2(explist = list(WT_40kb),
                  anchors = anchors,
                  shift = 0)
}
