% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PESCAn.R
\name{PESCAn_covert}
\alias{PESCAn_covert}
\title{PESCAn_covert}
\usage{
PESCAn_covert(experiment, bed, minComparables = 5, rmOutlier = F,
  minDist = 5e+06, size = 5e+05, add = 0)
}
\arguments{
\item{experiment}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{bed}{A bed-dataframe.}

\item{minComparables}{The minimal amount of bed-entries for a given chromosome. If this threshold is not reached, PE-SCAn will skip this chromosome.}

\item{rmOutlier}{to outlier-correction}

\item{minDist}{The minimal distance}

\item{size}{The amount of Hi-C bins to take into account (i.e. a score of 21 yield an output with 10 Hi-C bins up- and downstream of the anchor).}

\item{add}{Add constant value to bed-start and -end.}
}
\value{
A score-matrix.
}
\description{
From a ChIP bed file, a HiCpro matrix and a HiCpro bed file calculate a PE-scan like data structure. Run this per chromosome.
}
\examples{
# Run PE-SCAn on a bed of super-enhancers, using WT Hi-C data.
SE <- read.delim(superEnhancers.bed, header = F)
SE_vs_WT <- PESCAn(experiment = WT, bed = SE, size = 500e3)

# Do a circular perutation by adding 1Mb.
SE_vs_WT_perm <- PESCAn(experiment = WT, bed = SE, size = 500e3, add = 1e6)

# Plot using, for example, persp
persp(SE_vs_WT/SE_vs_WT_perm, phi = 30, theta = 30, col = 'skyblue')

}
