% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.chrom.comparison.R
\name{plot.chrom.comparison}
\alias{plot.chrom.comparison}
\title{plot.chrom.comparison}
\usage{
\method{plot}{chrom.comparison}(mat, color.fun = NULL, z.max = NULL)
}
\arguments{
\item{mat}{Output of chromosomeMatrix}

\item{color.fun}{Optional color-function}

\item{z.max}{Adjust the maximum value of the color-scale}
}
\value{
A plot of the matrix
}
\description{
Plot a genome-wide overview-matrix.
}

