% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation_score.R
\name{plot.insulation.dual}
\alias{plot.insulation.dual}
\title{plot.insulation.dual}
\usage{
\method{plot}{insulation.dual}(exp1, exp2, chrom, start, end, cut.off = 0,
  window.size = 21, local = T, delta = F)
}
\arguments{
\item{exp1}{A Hi-C experiment object: produced by construct.experiment().}

\item{exp2}{A Hi-C experiment object: produced by construct.experiment().}

\item{chrom}{Chromosome}

\item{start}{Start position of the region of interest}

\item{end}{End position of the region of interest}

\item{cut.off}{The cut.off for Hi-C scores}

\item{window.size}{The sliding square size}

\item{local}{Per chromosome?}

\item{delta}{plot the differential insulation scores}
}
\value{
A plot
}
\description{
Plot insulation-scores of two experiments
}

