% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation_score.R
\name{plot.insulation.single}
\alias{plot.insulation.single}
\title{plot.insulation.single}
\usage{
\method{plot}{insulation.single}(exp1, chrom, start, end, cut.off = 0,
  window.size = 21, local = T)
}
\arguments{
\item{exp1}{The Hi-C experiment object: produced by construct.experiment().}

\item{chrom}{Chromosome}

\item{start}{Start position of the region of interest}

\item{end}{End position of the region of interest}

\item{cut.off}{The cut.off for Hi-C scores}

\item{window.size}{The sliding square size}

\item{local}{Per chromosome?}
}
\value{
A plot
}
\description{
Plot a single insulation
}

