% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_plot.R
\name{plot_discovery}
\alias{plot_discovery}
\alias{plot.APA_discovery}
\alias{plot.CSCAn_discovery}
\alias{plot.PESCAn_discovery}
\alias{plot.ATA_discovery}
\alias{plot.ARA_discovery}
\alias{plot.CS_discovery}
\alias{plot.IS_discovery}
\alias{plot.DI_discovery}
\alias{plot.saddle_discovery}
\alias{plot.domainogram_discovery}
\alias{plot.RCP_discovery}
\alias{plot.virtual4C_discovery}
\alias{plot.IIT_discovery}
\title{Plot discoveries}
\usage{
\method{plot}{APA_discovery}(
  x,
  contrast = 1,
  colour_fun = NULL,
  colour_fun_contrast = NULL,
  colour_lim = NULL,
  colour_lim_contrast = NULL,
  ...
)

\method{plot}{CSCAn_discovery}(x, colour_fun = NULL, colour_lim = NULL, ...)

\method{plot}{PESCAn_discovery}(
  x,
  contrast = 1,
  colour_fun = NULL,
  colour_fun_contrast = NULL,
  colour_lim = NULL,
  colour_lim_contrast = NULL,
  ...
)

\method{plot}{ATA_discovery}(
  x,
  contrast = 1,
  colour_fun = NULL,
  colour_fun_contrast = NULL,
  colour_lim = NULL,
  colour_lim_contrast = NULL,
  ...
)

\method{plot}{ARA_discovery}(
  x,
  contrast = 1,
  colour_fun = NULL,
  colour_fun_contrast = NULL,
  colour_lim = NULL,
  colour_lim_contrast = NULL,
  ...
)

\method{plot}{CS_discovery}(x, chr = "chr1", start = NULL, end = NULL, contrast = NULL, ...)

\method{plot}{IS_discovery}(x, chr = "chr1", start = NULL, end = NULL, contrast = NULL, ...)

\method{plot}{DI_discovery}(x, chr = "chr1", start = NULL, end = NULL, contrast = NULL, ...)

\method{plot}{saddle_discovery}(
  x,
  contrast = 1,
  colour_fun = NULL,
  colour_fun_contrast = NULL,
  colour_lim = NULL,
  colour_lim_contrast = NULL,
  ...
)

\method{plot}{domainogram_discovery}(x, colour_fun = NULL, colour_lim = c(-1, 1), minimalist = FALSE, ...)

\method{plot}{RCP_discovery}(x, contrast = 1, metric = c("smooth"), ...)

\method{plot}{virtual4C_discovery}(x, censor_vp = TRUE, ...)

\method{plot}{IIT_discovery}(
  x,
  contrast = 1,
  censor_contrast = TRUE,
  geom = c("boxplot", "point"),
  ...
)
}
\arguments{
\item{x}{A \code{\link[GENOVA]{discovery}} object.}

\item{contrast}{An \code{integer} or \code{character} matching an experiment
(name) of length 1, specifying which sample should be used to make a
contrast with all other samples. Alternatively, set to \code{NULL} to not
plot contrast panels.}

\item{colour_fun, colour_fun_contrast}{A \code{function} that takes an integer
as input and returns colours as output, to colour raster-like data or the
constrast panels of raster-like data. Typically a function generated by
\code{\link[grDevices]{colorRampPalette}}.}

\item{colour_lim, colour_lim_contrast}{A \code{numeric} of length 2 indicating
the minimum and maximum values for which colours should be plotted.}

\item{...}{Not currently used for discovery plots.}

\item{chr}{A \code{character} vector of length 1 with the name of a
chromosome.}

\item{start}{An \code{integer} of length 1 with the start basepair of the
region to be plotted.}

\item{end}{An \code{integer} of length 1 with the ending basepair of the
region to be plotted.}

\item{minimalist}{\code{[domainogram]} A \code{logical} of length 1. If
\code{TRUE}, plot only the first experiment in the discovery and don't plot
any axis, titles or legends.}

\item{metric}{\code{[RCP]} Currently not in use.}

\item{censor_vp}{\code{[virtual_4C]} A \code{logical} of length 1 deciding
wether the values underneath the viewpoint should be excluded.}

\item{censor_contrast}{\code{[IIT]} A \code{logical} of length 1 deciding
wether the contrasting experiment itself should be censored (\code{TRUE})
or included (\code{FALSE}).}

\item{geom}{A \code{character} of length 1 indicating what geometry should be
plotted. Either \code{"boxplot"} or \code{"point"}.}
}
\value{
Nothing, but outputs a plot to the graphics device.
}
\description{
The base R plotting equivalent to
  \code{\link[GENOVA]{visualise}} for discovery objects.
}
\examples{
\dontrun{
plot(discovery)
}
}
