% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramids.R
\name{pyramidannotations}
\alias{pyramidannotations}
\alias{add_tads}
\alias{add_loops}
\alias{ggplot_add.mark_layer}
\title{Rotated pyramid annotations}
\usage{
add_tads(bed, colour = "#1faee3", ...)

add_loops(bedpe, colour = "#1faee3", shape = 1, ...)
}
\arguments{
\item{bed}{A BED-formatted \code{data.frame} with the following 3 columns:
\enumerate{ \item A \code{character} giving the chromosome names. \item An
\code{integer} with start positions. \item An \code{integer} with end
positions. }}

\item{colour, shape, ...}{Additional parameters to pass to
\code{ggplot2::layer()}.}

\item{bedpe}{A BEDPE-formatted \code{data.frame} with the following 6
columns: \enumerate{ \item A \code{character} giving the chromosome names
of the first coordinate. \item An \code{integer} giving the start positions
of the first coordinate. \item An \code{integer} giving the end positions
of the first coordinate. \item A \code{character} giving the chromosome
names of the second coordinate. \item An \code{integer} giving the start
positions of the second coordinate. \item An \code{integer} giving the end
positions of the second coordinate. }}
}
\value{
A \code{mark_layer} object.
}
\description{
Add annotations to a pyramid plot that also require the 45
  degree rotation.
}
\examples{
\dontrun{
p <- pyramid(exp, "chr2", 25e6, 30e6)
p + add_tads(tads)
p + add_loops(loops)
}
}
\keyword{internal}
