% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramids.R
\name{pyramidtracks}
\alias{pyramidtracks}
\alias{as_track}
\alias{add_bed_track}
\alias{add_bed_graph}
\alias{ggplot_add.track_layer}
\title{Adding tracks to a pyramid plot.}
\usage{
as_track(layer, size = 0.1, name, scale_y = NULL)

add_bed_track(bed, size = 0.1, name, ...)

add_bed_graph(bedgraph, size = 0.1, name, ...)
}
\arguments{
\item{layer}{A ggplot layer instance}

\item{size}{A \code{numeric} of length one noting the relative size to the
main facet panel.}

\item{name}{A \code{character} of length one title to display as facet strip
label.}

\item{scale_y}{A ggplot position scale with the y-aesthetic to control
parameters of the track's y-scale.}

\item{bed}{A BED-formatted \code{data.frame} with the following 3 columns:
\enumerate{ \item A \code{character} giving the chromosome names. \item An
\code{integer} with start positions. \item An \code{integer} with end
positions. }}

\item{...}{Additional arguments to pass to \code{ggplot2::layer()}.}

\item{bedgraph}{A \code{data.frame} with four columns noting the chromosome
name, start- and end-positions of genomic intervals, and a value to
display.}
}
\value{
A \code{track_layer} object.
}
\description{
These functions add new facets to a pyramid plot with additional
  annotations.
}
\details{
Currently, \code{as_track()} only accepts layers wherein the
  \code{data} argument is explicitly defined before evaluation. For example,
  \code{geom_point(aes(x = 1, y = 1))} is not an appropriate track, while
  \code{geom_point(aes(x + 1, y = y), data = data.frame(x = 1, y = 1))} is an
  appropriate track.
}
\examples{
\dontrun{
# Initialising a pyramid
p <- pyramid(exp, "chr2", 25e6, 30e6)

# Adding a bed annotation
p + add_bed_track(bed, fill = "blue", name = "bed")

# Adding a bedgraph annotation
p <- p + add_bed_graph(bedgraph, linetype = 2, name = "Some statistic")
p

# Adding custom annotation
# Use the name to display in the same panel as previous annotation
p + as_track(ggplot2::geom_vline(xintercept = 27.5e6),
             name = "Some statistic")
}
}
\keyword{internal}
