% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCP.region.R
\name{RCP.region}
\alias{RCP.region}
\title{Get a per-chromosome scaling plot}
\usage{
RCP.region(experimentList, bed, verbose = F)
}
\arguments{
\item{bed}{A BED-like dataframe of the regions of interest}

\item{verbose}{Produces a progress-indication.}

\item{experiment}{List of experiment-objects from `construct.experiment()`.}
}
\value{
A data_frame with distance-bin and probabilities.
}
\description{
Produce a dataframe with the probabilities of contacts in a set of distance-bins.
Bins are created on a log scale, which leads to equal amounts of datapoints per bin.
}
