% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCP.R
\name{RCPlfc}
\alias{RCPlfc}
\title{RCP log2 foldchange

 RAW RCP in, lfc out}
\usage{
RCPlfc(dt, contrast, breaks)
}
\arguments{
\item{dt}{a data.table of rcp}

\item{contrast}{the name of the contrast-sample}

\item{breaks}{a set of numbers to use as intervals}
}
\value{
a data.table with the log2 fold changes compared to the `contrast`.
}
\description{
RCP log2 foldchange

 RAW RCP in, lfc out
}
\keyword{internal}
