% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bedpe.R
\name{read_bedpe}
\alias{read_bedpe}
\title{Load, parse and check BEDPE.}
\usage{
read_bedpe(loop.bed, header = T)
}
\arguments{
\item{loop.bed}{Path to bedpe file.}

\item{header}{Does the file come with a header?}
}
\value{
A data.frame with the bedpe-values.
}
\description{
This function loads a BEDPE file. Columns 2 and 4 and 5 and 6 will be sorted numerically.
If only a BED-file is present, duplicate BED-columns 1,2,3 as: 1,2,2,1,3,3.
}
