% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_matrix_lookup.R
\name{rep_mat_lookup}
\alias{rep_mat_lookup}
\title{Repeated Hi-C matrix lookup}
\usage{
rep_mat_lookup(explist, anchors, rel_pos, shift = 0,
  outlier_filter = c(0, 1), raw = FALSE)
}
\arguments{
\item{explist}{Either a single GENOVA experiment or a list of GENOVA
experiments.}

\item{anchors}{A \code{matrix} with two columns containing pre-computed
anchor indices.}

\item{rel_pos}{An \code{integer} vector indicating relative positions in
bins, indicating ranges around anchors to lookup.}

\item{shift}{An \code{integer} of length 1 indicating how many basepairs the
anchors should be shifted. Essentially performs circular permutation of
\code{size} for a reasonable estimate of background. The argument is
ignored when \code{shift <= 0}.}

\item{outlier_filter}{A \code{numeric} of length 2 between \code{[0-1]}
indicating quantiles of data to be used as thresholds. Data outside these
thresholds are set to the nearest threshold. Setting this to \code{c(0, 1)}
performs no outlier correction.}

\item{raw}{A \code{logical} of length 1: should the raw array underlying
the summary matrices be returned in the output?}
}
\value{
A \code{list} of the same length as \code{explist} wherein list
  elements contain the results of the repeated lookup per experiment.
}
\description{
Analyses like aggregrate peak analysis (APA) and paired-end spatial chromatin
analysis (PE-SCAn) require to look up regions of a Hi-C matrix repeatedly.
This function acts as a common method for the lookup and summary of such
repeated lookups for these analyses.
}
\seealso{
\code{\link[GENOVA]{APA2}} and \code{\link[GENOVA]{PESCAn2}}
}
