% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale a vector.}
\usage{
rescale(x, newrange = range(x))
}
\arguments{
\item{x}{A vector which needs to be resized.}

\item{newrange}{A vector of the wanted length.}
}
\value{
A rescaled vector.
}
\description{
Rescale a vector.
}
