% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize.mat.R
\name{resize.mat}
\alias{resize.mat}
\title{Rescale a matrix.}
\usage{
resize.mat(mat, ndim = dim(mat))
}
\arguments{
\item{mat}{A vector which needs to be resized.}

\item{ndim}{Dimentions of the wanted matrix: can be obtained by dim(matrixOfChoice).}
}
\value{
A rescaled vector.
}
\description{
Rescale a matrix.
}
