% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_resolution.R
\name{resolution}
\alias{resolution}
\title{Get the resolution of GENOVA objects}
\usage{
resolution(x)
}
\arguments{
\item{x}{An object from the GENOVA package}
}
\value{
A \code{integer} with the resolution
}
\description{
Get the resolution of GENOVA objects
}
\examples{
\dontrun{
resolution(contacts_object)
resolution(discovery_object)
}
}
