% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddle.R
\name{saddle}
\alias{saddle}
\title{Compartment versus compartment scores}
\usage{
saddle(explist, CS_discovery, bins = 10L, dist_thres = c(-Inf, Inf))
}
\arguments{
\item{explist}{Either a single GENOVA '\code{contacts}' object or a list of
GENOVA '\code{contacts}' objects.}

\item{CS_discovery}{A signed '\code{CS_discovery}' produced by the
\code{compartment_score} function.}

\item{bins}{An \code{integer} of length 1 setting the number of quantiles the
comparment score should be divided into.}

\item{dist_thres}{A \code{numeric} of length two noting the lower and upper 
limit of distances in basepairs to consider. Defaults to 
\code{c(-Inf, Inf)} to include all distances.}
}
\value{
A \code{saddle_discovery} object with 1 element:

\itemize{\item\strong{\code{saddle}}, a \code{data.table} with the following columns: 
\describe{
\item{\code{exp}}{A \code{character} with the sample names from the 
'\code{explist}' argument.}
\item{\code{chr}}{A \code{character} with the chromosome names and arms (p or q).}
\item{\code{q1}}{An \code{integer} giving the first comparment score quantile bin.
Lower values indicate smaller compartment scores than higher values.}
\item{\code{q2}}{An \code{integer} giving the second quantile bin.}
\item{\code{mean}}{A \code{numeric} with the average observed over expected values
at the indicated quantile bins.}
}}
}
\description{
Produces a compartment score quantile versus quantile analysis in which
observed over expected contacts are compared between the different quantiles
of the compartment scores.
}
\details{
Per chromosome arm, compartment scores are divided in quantile bins.
  Subsequently, the average observed over expected score is calculated for
  every pairwise combination of bins. The observed over expected score is
  calculated as the contacts divided by the average of contacts for the same
  distance from the diagonal.

  Choosing a '\code{bins}' of \code{5} will produce results similar to
  Flyamer \emph{et al}. (2007), while setting '\code{bins}' to \code{100}
  will produce results similar to Bonev \emph{et al}. (2017)
}
\section{Resolution recommendation}{
 The resolution of the
  '\code{CS_discovery}' object. Either print the object to see the resolution
  or use \code{attr(CS_discovery, "resolution")} for programmatic access.
}

\examples{
\dontrun{
exps <- list(WT_100kb, KO_100kb)

# Computing signed compartment scores
cs <- comparment_score(exps, bed = H3K4me1_peaks)

# Saddle analysis
sadl <- saddle(exps, cs)

# Visualising results
visualise(sadl)
}
}
