% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddle.R
\name{saddleBins}
\alias{saddleBins}
\title{Compute compartment-v-compartment scores}
\usage{
saddleBins(exp, ChIP = NULL, chromsToUse = NULL, nBins = 5, verbose = T)
}
\arguments{
\item{exp}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{ChIP}{BED-dataframe containing active sites (e.g. H3K27ac-peaks).}

\item{chromsToUse}{Do the computation for a subset of chromosomes.}

\item{nBins}{The number of bins to split the compartment-score. A binsize of 5 will produce a plot similar to FLyamer et al. (2017), while 100 will produce a plot similar to Bonev et al. (2017).}

\item{verbose}{Produces a progress-indication.}
}
\value{
A log2(O/E) matrix and a DF of compartment-scores.
}
\description{
Splits range of compartment-scores in nBins bins.
Generates an average O/E contact-score for each compartment-score bin.
Sorts matrix on compartment-score.
}
