% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_visualise.R
\name{scale_altfill_continuous}
\alias{scale_altfill_continuous}
\title{scale_altfill_continuous
Makes sure no errors are returned when visualise(..., raw = TRUE)
Unfortunately has to be exported, but is better than throwing errors}
\usage{
scale_altfill_continuous(...)
}
\arguments{
\item{...}{whatever.}
}
\value{
a thing
}
\description{
scale_altfill_continuous
Makes sure no errors are returned when visualise(..., raw = TRUE)
Unfortunately has to be exported, but is better than throwing errors
}
