% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.sub.2D.R
\name{select.sub.2D}
\alias{select.sub.2D}
\title{A quick lookup of square region from a HiC matrix.}
\usage{
select.sub.2D(data, X, Y)
}
\arguments{
\item{data}{HiC-pro matrix}

\item{X}{The X index-range.}

\item{Y}{The Y index-range.}
}
\value{
A data.table with normalised counts.
}
\description{
Extracts a symmetric matrix. *X* and *Y* renote the indexes to extracts.
}
