% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.sub.R
\name{select.sub}
\alias{select.sub}
\title{A quick lookup of square region from a HiC matrix.}
\usage{
select.sub(data, start, end)
}
\arguments{
\item{data}{HiC-pro matrix}

\item{start}{Start of index-range.}

\item{end}{End of index-range.}
}
\value{
A data.table with normalised counts.
}
\description{
Extracts a symmetric matrix around the diagonal from *start* to *stop*.
}
