% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.subset.R
\name{select.subset}
\alias{select.subset}
\title{Get matrix from a BED-like entry.}
\usage{
select.subset(r1, chrom, start, end, bed)
}
\arguments{
\item{r1}{A HiC-pro matrix.}

\item{chrom}{Chromosome.}

\item{start}{Start position in bp.}

\item{end}{End position in bp.}

\item{bed}{A HiC-pro index.}
}
\value{
A data.table with normalised counts.
}
\description{
Extracts a symmetric matrix around the diagonal from *start* to *stop* in chromosome *chrom*.
}
