% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_subset}
\alias{select_subset}
\title{Get a matrix from a BED-like entry}
\usage{
select_subset(exp, chrom, start, end)
}
\arguments{
\item{exp}{The \code{contacts} objects produced by
\code{construct.experiment()}}

\item{chrom}{A \code{character} of length 1: the chromosome.}

\item{start}{An \code{integer} of length 1 noting the start position in bp.}

\item{end}{An \code{integer} of length 1 noting the end position in bp.}
}
\value{
A list with the \code{X} and \code{Y} coordinates and a \code{matrix
  Z} containing the contacts at these coordinates.
}
\description{
Extracts a square symmetric matrix around the diagonal from a \code{contacts}
object.
}
\examples{
\dontrun{
# Get the TP53-locus in an experiment mapped to hg19
mat <- select_subset(WT, "chr17", 75e5, 76e5)

# Plot the region
image(mat)
}
}
