% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked.TAD.R
\name{stacked.TAD}
\alias{stacked.TAD}
\title{Get a z-stack matrix of TADs}
\usage{
stacked.TAD(experiment, tad.bed, smallTreshold = 225000, verbose = F)
}
\arguments{
\item{experiment}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{tad.bed}{Bedpe file containing the TAD positions: produced by read.bedpe(). Please make a bedpe, by copying tad-bed columns to columns 4,5 and 6!}

\item{smallTreshold}{The minimal size of loops. Too small loops lead to messy plots.}

\item{verbose}{Produces a progress-indication.}
}
\value{
A matrix containing the Z-stack scores.
}
\description{
Extracts matrices from a BED-like structure and resizes them, which leads to all start- and end-position of TADs overlapping.
Sums over all matrices to produce a single Z-stack matrix, which is normalised to 100 loops.
}

