% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stacked_Elzo.R, R/stackR.R
\name{stackR}
\alias{stackR}
\title{Get a z-stack matrix of TADs}
\usage{
stackR(experiment, tad.bed, smallTreshold = 225000, verbose = F,
  saveRaw = T, saveRawList = T, outlierCutOff = 50)

stackR(experiment, tad.bed, smallTreshold = 225000, verbose = F,
  saveRaw = T, saveRawList = T, outlierCutOff = 50)
}
\arguments{
\item{experiment}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{tad.bed}{Data.frame from a Bed file containing the TAD positions.}

\item{smallTreshold}{The minimal size of loops. Too small loops lead to messy plots.}

\item{verbose}{Produces a progress-indication.}

\item{saveRaw}{Logical: True will output an aditional matrix without outlier-correction.}

\item{saveRawList}{Logical: True will output the raw matrices per TAD.}

\item{outlierCutOff}{The severity of outliers: roughly translates to the amount of MADs above the median.}

\item{experiment}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{tad.bed}{Data.frame from a Bed file containing the TAD positions.}

\item{smallTreshold}{The minimal size of loops. Too small loops lead to messy plots.}

\item{verbose}{Produces a progress-indication.}

\item{saveRawList}{Logical: True will output the raw matrices per TAD.}

\item{saveRaw}{Logical: True will output an aditional matrix without outlier-correction.}

\item{outlierCutOff}{The severity of outliers: roughly translates to the amount of MADs above the median.}
}
\value{
A list containing a matrix with the Z-stack scores (`STACK`). Optionally, it can contain the raw matrices per TAD (`STACK.list`) and a matrix without the outlier-correction (`STACK.raw`).

A list containing a matrix with the Z-stack scores (`STACK`). Optionally, it can contain the raw matrices per TAD (`STACK.list`) and a matrix without the outlier-correction (`STACK.raw`).
}
\description{
Extracts matrices from a BED-like structure and resizes them, which leads to all start- and end-position of TADs overlapping.
Sums over all matrices to produce a single Z-stack matrix, which is normalised to 100 loops.

Extracts matrices from a BED-like structure and resizes them, which leads to all start- and end-position of TADs overlapping.
Sums over all matrices to produce a single Z-stack matrix, which is normalised to 100 loops.
}

