% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_contacts.R
\name{subset.contacts}
\alias{subset.contacts}
\title{subset}
\usage{
\method{subset}{contacts}(x, chrom = NULL, start = 0, end = Inf, ...)
}
\arguments{
\item{x}{A \code{contacts}-object}

\item{chrom}{A vector of chromosomes to subset}

\item{start}{A single postion in bp.}

\item{end}{A single posttion in bp.}

\item{...}{Currently not in use.}
}
\description{
subset
}
\examples{
\dontrun{
WT_chr1 <- subset(WT_new, chrom = 'chr1')
WT_acrocentric <- subset(WT_new, chrom = c('chr13',
                                           'chr14',
                                           'chr15',
                                           'chr21',
                                           'chr22'))
}
}
