% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_matrix_lookup.R
\name{summarise_lookup}
\alias{summarise_lookup}
\title{Summarise the results of a repeated matrix lookup}
\usage{
summarise_lookup(array, outlier_filter = c(0, 1), keep = NULL, group = NULL)
}
\arguments{
\item{array}{A threedimensional \code{array}.}

\item{outlier_filter}{A \code{numeric} of length 2 between \code{[0-1]}
indicating quantiles of data to be used as thresholds. Data outside these
thresholds are set to the nearest threshold. Setting this to \code{c(0, 1)}
performs no outlier correction.}

\item{keep}{A \code{logical} vector of length \code{dim(array)[1]} indicated
which slices to keep. If \code{NULL}, independently filters out
all-\code{NA} slices.}
}
\value{
A \code{list} of length 2 with a \code{matrix} giving mean values of
  \code{array} along the first dimension and the \code{keep} as used in the
  calculation of the means.
}
\description{
This function either filters out all the array slices that are solely
composed of \code{NA}s or those specified. It sets all other \code{NA}s to
\code{0} and then takes the mean along the first dimension of a
threedimensional array.
}
\seealso{
\code{\link[GENOVA]{rep_mat_lookup}}
}
\keyword{internal}
