% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_plot.R
\name{trans.compartment.plot}
\alias{trans.compartment.plot}
\title{Draw interchromosomal interaction heatmap for a chromosome (arm) with corresponding (cis) compartment scores}
\usage{
trans.compartment.plot(exp, chrom1, arm1 = "p", chrom2, arm2 = "p",
  zlim = 20, invert = c(F, F), color.scheme = "fall", cs.lim = NULL)
}
\arguments{
\item{exp}{GENOVA data object}

\item{chrom1, chrom2}{chromosome combination that should be drawn}

\item{arm1, arm2}{which chromosome arm: 'p' or 'q' (for acrocentric this should be set to 'q')}

\item{zlim}{maximum value for the heatmap}

\item{invert}{whether the compartment score should be inverted, this should be a vector with a logical value for each chromosome}

\item{color.scheme}{color scheme that should be used, defaults to fall, other values result in white-red gradient}

\item{cs.lim}{y-axis limit for the compart score, if unset (NULL) will default to the maximum absolute value}
}
\description{
Draw interchromosomal interaction heatmap for a chromosome (arm) with corresponding (cis) compartment scores
}
