% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixplot_trans.R
\name{trans_matrixplot}
\alias{trans_matrixplot}
\title{Trans matrix plot}
\usage{
trans_matrixplot(
  exp1,
  exp2 = NULL,
  chrom_up,
  start_up,
  end_up,
  chrom_down,
  start_down,
  end_down,
  colour_lim = NULL,
  rasterise = FALSE,
  colour_bar = FALSE
)
}
\arguments{
\item{exp1}{A GENOVA \code{contacts} object.}

\item{exp2}{A GENOVA \code{contacts} object compatible with \code{exp1} 
(optional) or \code{NULL}. If not \code{NULL} the \code{exp2} argument will 
be subtracted from the \code{exp1} data.}

\item{chrom_up, chrom_down}{One of the following: \itemize{
  \item{A \code{character} of length 1 giving the chromosome name of the 
  region of interest}
  \item{A 3-column, 1-row \code{data.frame} in BED-format}
  \item{A single \code{character} describing a locus in UCSC notation, e.g.
  \code{"chr1:30,000,000-40,000,000"}.}
} The latter two options automatically provide the \code{start_*} and 
  \code{end_*} arguments too.}

\item{start_up, start_down}{An \code{integer} with the start position of 
the regions of interest.}

\item{end_up, end_down}{An \code{integer} with the end position of the regions
of interest.}

\item{colour_lim}{A \code{numeric} of length two giving the minimum and
maximum values in contacts. Out of bounds values will be set to the nearest
limits.}

\item{rasterise}{Set to \code{TRUE} to use a bitmap raster instead of 
polygons.}

\item{colour_bar}{A \code{logical} of length 1, indicating whether a
colour-bar legend should be drawn at the right.}
}
\value{
No values are returned but a plot is outputted to the graphics 
  device.
}
\description{
This function plots a matrix of contacts data for the intersection of two
regions. Using two regions on different chromosomes plots trans data.
}
\section{Resolution recommendation}{
 A resolution in the ballpark of 
  \code{(end - start) / 500}.
}

\examples{
\dontrun{
trans_matrixplot(exp, 
                 chrom_up = "chr1:30,000,000-40,000,000",
                 chrom_down = "chr2:50,000,000-60,000,000")
}
}
\seealso{
Other matrix plots: 
\code{\link{compartment_matrixplot}()},
\code{\link{hic_matrixplot}()},
\code{\link{insulation_matrixplot}()},
\code{\link{pyramid}()}
}
\concept{matrix plots}
