% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual.4C.R
\name{virtual.4C}
\alias{virtual.4C}
\title{Get an average virtual 4C of 2D-regions (e.g. loops).}
\usage{
virtual.4C(experiment, loop.bed, smallTreshold = 225000, verbose = F)
}
\arguments{
\item{experiment}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{loop.bed}{Bedpe file containing the loop positions: produced by read.bedpe().}

\item{smallTreshold}{The minimal size of loops. Too small loops lead to messy plots.}

\item{verbose}{Produces a progress-indication.}
}
\value{
A vector containing the y-axis scores per rescaled bin.
}
\description{
Extract matrices around a defined set of pixels, like possible loops.
Matrices are rescaled, so that all viewpoints and anchors are overlapping.
}
