% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_4C.R
\name{virtual_4C}
\alias{virtual_4C}
\title{Get an average virtual 4C of 2D-regions (e.g. loops).}
\usage{
virtual_4C(exp, viewpoint, xlim = NULL)
}
\arguments{
\item{exp}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{viewpoint}{The viewpoint}

\item{xlim}{A vector of two with the flanking basepairs up- and downstream 
of the viewpoint, resp.}
}
\value{
A virtual4C_discovery object.
}
\description{
Extract matrices around a defined region a.k.a. the viewpoint
}
