% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.APA.base.R
\name{visualise.APA.base}
\alias{visualise.APA.base}
\title{visualise.APA.ggplot}
\usage{
visualise.APA.base(APAlist, title, focus = 1, color.fun1 = NULL,
  color.fun2 = NULL, abs.max = NULL, rel.max = NULL, ...)
}
\arguments{
\item{APAlist}{A list of results from the APA-function.}

\item{title}{Title-text to plot.}

\item{focus}{Which sample will be the to-compare sample in the differential row?}

\item{color.fun1}{Optional color function for the first row}

\item{color.fun2}{Optional color function for the second row}

\item{abs.max}{Optional Z-max of the first row}

\item{rel.max}{Optional Z-max of the second row}
}
\value{
A plot
}
\description{
Plot the APA-results and the differential results.
}
\details{
By substracting the values from each sample with the values from the \code{focus}-sample, we generate the differentials.
A positive value in the differential plots thus means an enrichment in that sample versus the \code{focus}-sample.
}
\author{
Elzo de Wit, \email{e.d.wit@nki.nl}
}
