% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.APA.ggplot.R
\name{visualise.APA.ggplot}
\alias{visualise.APA.ggplot}
\title{Plot the APA-results}
\usage{
visualise.APA.ggplot(APAlist, title, zCoef = 1, focus = 1, ...)
}
\arguments{
\item{APAlist}{A list of results from `APA`.}

\item{title}{Text to plot}

\item{zCoef}{Coefficient to change the zlims}

\item{Focus}{Wich sample does need to be the to-compare sample?}
}
\value{
A grid object, containing two ggplot-objects.
}
\description{
Plot the APA-results
}

