% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.APA.base.R
\name{visualise.APA}
\alias{visualise.APA}
\title{Plot the APA-results}
\usage{
visualise.APA(APAlist, title, focus = 1, color.fun1 = NULL,
  color.fun2 = NULL, abs.max = NULL, rel.max = NULL, ...)
}
\arguments{
\item{APAlist}{A list of results from `APA`.}

\item{title}{Text to plot}

\item{color.fun1}{Optional color function for the first row}

\item{color.fun2}{Optional color function for the second row}

\item{abs.max}{Optional Z-max of the first row}

\item{rel.max}{Optional Z-max of the second row}

\item{Focus}{Wich sample does need to be the to-compare sample?}
}
\value{
A grid object, containing two ggplot-objects.
}
\description{
Plot the APA-results
}

