% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddle.R
\name{visualise.compartmentStrength}
\alias{visualise.compartmentStrength}
\title{Plot compartment-strengths}
\usage{
visualise.compartmentStrength(SBoutList)
}
\arguments{
\item{SBoutList}{A list of outputs trom saddleBins.}
}
\value{
A plot and an invisible underlying dataframe.
}
\description{
Takes the input from saddleBins and produces a boxplot of compartment-strenghts.
}
