% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PESCAn.R
\name{visualise.PESCAn.ggplot}
\alias{visualise.PESCAn.ggplot}
\title{Plot the PE-SCAn-results}
\usage{
visualise.PESCAn.ggplot(PESCAnlist, resolution, title = "PE-SCAn",
  zTop = NULL, zBottom = NULL, focus = 1, smooth = F, ...)
}
\arguments{
\item{PESCAnlist}{A list of results from `PESCAn`.}

\item{title}{Text to plot}

\item{zTop}{The min and max colorscale-values for the first row of plots.}

\item{zBottom}{The min and max colorscale-values for the first row of plots.}

\item{Focus}{Which sample does need to be the to-compare sample?}
}
\value{
A grid object, containing two ggplot-objects.
}
\description{
Plot the PE-SCAn-results
}
