% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.RCP.ggplot.R
\name{visualise.RCP.ggplot}
\alias{visualise.RCP.ggplot}
\title{Plot the RCP results}
\usage{
visualise.RCP.ggplot(RCPdata, smooth = F, combine = T, lineWidth = 1,
  pointWidth = 0.5)
}
\arguments{
\item{RCPdata}{Output of `RCP`}

\item{smooth}{Plot a lowess-smoothed line?}

\item{combine}{Combine all chromosomes?}

\item{lineWidth}{Width of SEM-lines}

\item{pointWidth}{Width of points}
}
\value{
A ggplot-object
}
\description{
Produces a line-plot per chromosome.
}
