% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods_visualise.R
\name{visualise}
\alias{visualise}
\alias{visualise.default}
\alias{visualise.APA_discovery}
\alias{visualise.PESCAn_discovery}
\alias{visualise.ATA_discovery}
\alias{visualise.ARA_discovery}
\title{Visualise discoveries}
\usage{
visualise(discovery, ...)

\method{visualise}{APA_discovery}(discovery, contrast = 1,
  metric = c("diff", "lfc"), raw = FALSE)

\method{visualise}{PESCAn_discovery}(discovery, contrast = 1,
  metric = c("diff", "lfc"), mode = c("obsexp", "signal"),
  raw = FALSE)

\method{visualise}{ATA_discovery}(discovery, contrast = 1,
  metric = c("diff", "lfc"), raw = FALSE)

\method{visualise}{ARA_discovery}(discovery, contrast = 1,
  metric = c("diff", "lfc"), raw = FALSE)
}
\arguments{
\item{discovery}{A \code{discovery} object as returned by GENOVA analysis
functions.}

\item{contrast}{An \code{integer} or \code{character} matching an experiment
(name) of length 1, specifying which sample should be used to make a
contrast with all other samples. Alternatively, set to \code{NULL} to not
plot contrast panels.}

\item{metric}{A \code{character} of length 1: what contrast metric should be
used? \code{"diff"} for difference by subtraction or \code{"lfc"} for
\ifelse{html}{\out{p_<sub>x</sub>(1)=Pr(y=1|x)}}{\eqn{p_x(1)=Pr(y=1|x)}}
fold changes.}

\item{raw}{A \code{logical} of length 1: should a bare bones plot be
returned?}

\item{mode}{(PESCAn_discovery only) What result slot should be used for
visualisation? \code{"obsexp"} for the observed over expected metric and
\code{"signal"} for mean contacts at unshifted anchors.}
}
\description{
Plot the results of \code{discovery} objects. By default
  contrasts one sample with the others.
}
\details{
The \code{"diff"} \code{metric} value creates contrast panels by
  subtracting the values of each sample by the values of the sample indicated
  by the '\code{contrast}' argument. The \code{"lfc"} \code{metric} value
  creates contrast panels by dividing every samples' values by the sample
  indicated by the '\code{contrast}' argument and then taking the base 2
  logarithm of that division.

  The '\code{raw = TRUE}' argument allows custimisation of the plots. The
  returned plot will not have position- or fill-scales and no theme settings,
  which can be set freely afterwards to match personal aesthetic tastes. When
  '\code{raw = TRUE}' and '\code{subtract}' is not \code{NULL}, the fill
  scale of the contrast panels can be manipulated by setting the
  '\code{aesthetics = "altfill"}' inside ggplot2's fill scale functions.
}
\note{
For \code{ATA_discovery} and \code{ARA_discovery} objects which include
  the matrix's diagonal the upper limit for the contacts fill scale is set to
  the 95th percentile of the data to increase the dynamic range of colours.
}
\examples{
# APA
apa <- APA(list(WT = WT_40kb, KO = KO_40kb), loops)
visualise(apa)

# PE-SCAn
pescan <- PESCAn(list(WT = WT_40kb, KO = KO_40kb), super_enhancers)
visualise(pescan)

# To plot PE-SCAn without background correction
visualise(pescan, mode = "signal")

# ATA
ata <- APA(list(WT = WT_10kb, KO = KO_10kb), tads)
visualise(ata)

# ARA
ara <- ARA(list(WT = WT_20kb, KO = KO_20kb), ctcf_sites)
visualise(ara)

# Handling 'raw' plots
visualise(pescan, raw = TRUE) +
  ggplot2::scale_fill_gradient(aesthetics = "altfill")
}
