% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.stacked.TAD.ggplot.R
\name{visualise.stacked.TAD.ggplot}
\alias{visualise.stacked.TAD.ggplot}
\title{Plot the stacked TAD results.}
\usage{
visualise.stacked.TAD.ggplot(stackedlist, title, focus = 1)
}
\arguments{
\item{stackedlist}{List of results from `stacked.TAD`.}

\item{title}{Text to plot}

\item{Focus}{Wich sample does need to be the to-compare sample?}
}
\value{
A grid object, containing two ggplot-objects.
}
\description{
Plot the stacked TAD results.
}

