% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.stackR.ggplot.R, R/visualise.stacked.TAD.ggplot.R
\name{visualise.stackR.ggplot}
\alias{visualise.stackR.ggplot}
\title{Plot the stacked TAD results.}
\usage{
visualise.stackR.ggplot(stackedlist, title, focus = 1, zCoef = 1, ..)

visualise.stackR.ggplot(stackedlist, title, focus = 1, zCoef = 1, ..)
}
\arguments{
\item{stackedlist}{Named list of results from `stacked.TAD`.}

\item{title}{Text to plot}

\item{zCoef}{Coefficient to change the zlims above}

\item{Focus}{Wich sample does need to be the to-compare sample?}

\item{z2Coef}{Coefficient to change the zlims below}

\item{stackedlist}{List of results from `stacked.TAD`.}

\item{title}{Text to plot}

\item{Focus}{Wich sample does need to be the to-compare sample?}

\item{zCoef}{Coefficient to change the zlims}
}
\value{
A grid object, containing two ggplot-objects.

A grid object, containing two ggplot-objects.
}
\description{
Plot the stacked TAD results.

Plot the stacked TAD results.
}

