% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.stackR2.ggplot.R
\name{visualise.stackR_V2.ggplot}
\alias{visualise.stackR_V2.ggplot}
\title{Plot the stacked TAD results.}
\usage{
visualise.stackR_V2.ggplot(stackedlist, title, focus = 1, zCoef = 1,
  z2Coef = 1)
}
\arguments{
\item{stackedlist}{Named list of results from `stacked.TAD`.}

\item{title}{Text to plot}

\item{zCoef}{Coefficient to change the zlims above}

\item{z2Coef}{Coefficient to change the zlims below}

\item{Focus}{Wich sample does need to be the to-compare sample?}
}
\value{
A grid object, containing two ggplot-objects.
}
\description{
Plot the stacked TAD results.
}
