% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.virtual.4C.R
\name{visualise.virtual.4C}
\alias{visualise.virtual.4C}
\title{Plot the virtual 4C results.}
\usage{
visualise.virtual.4C(data)
}
\arguments{
\item{data}{Output of `virtual.4C`.}
}
\value{
A plot.
}
\description{
Plot the virtual 4C results.
}
