% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzvisualise.virtual.4C.R
\name{visualise_virtual_4C}
\alias{visualise_virtual_4C}
\title{Plot the virtual 4C results.}
\usage{
visualise_virtual_4C(data)
}
\arguments{
\item{data}{Output of `virtual.4C`.}
}
\value{
A plot.
}
\description{
Plot the virtual 4C results.
}
