% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMLA.R
\name{ARA}
\alias{ARA}
\title{Aggregate Region Analysis}
\usage{
ARA(explist, bed, shift = 1e+06, size_bin = 21, size_bp = NULL,
  outlier_filter = c(0, 1), anchors = NULL, raw = FALSE)
}
\arguments{
\item{explist}{Either a single GENOVA experiment or a list of GENOVA
experiments.}

\item{bed}{A \code{data.frame} with 3 columns in BED format, containing the
regions to anchor in pairwise manner. Entries wherein the second column is
larger than the third column are considered in the reverse direction.}

\item{shift}{An \code{integer} of length 1 indicating how many basepairs the
anchors should be shifted. Essentially performs circular permutation of
\code{size} for a reasonable estimate of background. The argument is
ignored when \code{shift <= 0}.}

\item{size_bin}{The size of the lookup regions in bins (i.e. a score of 21
yields an output with 10 Hi-C bins both up- and downstream of the anchor).}

\item{size_bp}{Alternative parametrisation for the lookup regions, expressed
in basepairs. Not used when the argument \code{size_bin} is set.}

\item{outlier_filter}{A \code{numeric} of length 2 between \code{[0-1]}
indicating quantiles of data to be used as thresholds. Data outside these
thresholds are set to the nearest threshold. Setting this to \code{c(0, 1)}
performs no outlier correction.}

\item{anchors}{(Optional) A \code{matrix} with two columns containing
pre-computed anchor indices. If this is set, skips calculation of anchor
indices and uses these instead.}

\item{raw}{A \code{logical} of length 1: should the raw array underlying the
summary matrices be returned in the output?}
}
\value{
An \code{ARA_discovery} object with the ARA results.
}
\description{
Extracts Hi-C matrices centered around regions and averages the results for
all regions.
}
\details{
By default, \code{ARA} also calculates the results for shifted
  anchors and normalises the \code{"obsexp"} slot by off-diagonal bands.

  The '\code{bed}' argument can take in oriented entries, wherein entries
  with a start site larger than the end site are considered to be in the
  reverse direction. The reverse sites are flipped during analysis, so the
  orientation is the same as in the forward sites.
}
\examples{
\dontrun{
# Typical usage
ara <- ARA(list(WT_20kb, KO_20kb), ctcf_sites)

# Alternative usage with pre-calculated anchors
anchors <- anchors_ARA(WT_20kb$ABS, ctcf_sites)
ara <- ARA(list(WT_20kb, KO_20kb))

# Visualisation
visualise(ara)
}
}
\seealso{
The \code{\link[GENOVA]{discovery}} class.
  \code{\link[GENOVA]{visualise}} for visualisation of the results.
}
