% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA.R
\name{ATA}
\alias{ATA}
\title{Get a z-stack matrix of TADs}
\usage{
ATA(experiment, tad.bed, smallTreshold = 225000, verbose = F, saveRaw = T,
  saveRawList = T, outlierCutOff = 40)
}
\arguments{
\item{experiment}{The Hi-C experiment object of a sample: produced by \code{construct.experiment()}.}

\item{tad.bed}{Data.frame from a Bed file containing the TAD positions.}

\item{smallTreshold}{The minimal size of loops. Too small loops lead to messy plots.}

\item{verbose}{Produces a progress-indication.}

\item{saveRaw}{Logical: True will output an aditional matrix without outlier-correction.}

\item{saveRawList}{Logical: True will output the raw matrices per TAD.}

\item{outlierCutOff}{The severity of outliers: roughly translates to the amount of MADs above the median.}
}
\value{
\item{STACK}{A list containing a matrix with the Z-stack scores}

\item{STACK.list}{Optional: raw matrices per TAD}

\item{STACK.raw}{Optional: a matrix without the outlier-correction}
}
\description{
Extracts matrices from a BED-like structure and resizes them, which leads to all start- and end-position of TADs overlapping.
Sums over all matrices to produce a single Z-stack matrix, which is normalised to 100 loops.
}
\examples{
ATA_results_of_RAOetal <- ATA(experiment = Rao_20k,tad.bed = TADs)
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
