% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMLA.R
\name{ATA}
\alias{ATA}
\title{Aggregate TAD analysis}
\usage{
ATA(explist, bed, dist_thres = c(225000, Inf), size = 100L,
  outlier_filter = c(0, 0.995), anchors = NULL, raw = TRUE)
}
\arguments{
\item{explist}{Either a single GENOVA experiment or a list of GENOVA
experiments.}

\item{bed}{A \code{data.frame} with 3 columns in BED format, containing the
TAD boundary positions per row.}

\item{dist_thres}{An \code{integer} vector of length 2 indicating the minimum
and maximum sizes of TADs to include in basepairs.}

\item{size}{A code \code{integer} vector of length 1 noting the dimensions of
the output.}

\item{outlier_filter}{A \code{numeric} of length 2 between \code{[0-1]}
indicating quantiles of data to be used as thresholds. Data outside these
thresholds are set to the nearest threshold. Setting this to \code{c(0, 1)}
performs no outlier correction.}

\item{anchors}{(Optional) A \code{matrix} with two columns containing
pre-computed anchor indices. If this is set, skips calculation of anchor
indices and uses these instead.}

\item{raw}{A \code{logical} of length 1: should the raw array underlying the
summary matrices be returned in the output?}
}
\value{
An \code{ATA_discovery} object with the ATA results.
}
\description{
Extracts Hi-C matrices around TADs, resizes these to a uniform size and
averages the results for all TADs.
}
\seealso{
The \code{\link[GENOVA]{discovery}} class.
  \code{\link[GENOVA]{visualise}} for visualisation of the results.
}
