% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA.R
\name{ATA}
\alias{ATA}
\title{Aggegrate TAD Analysis}
\usage{
ATA(experiment, tad.bed, smallTreshold = 225000, rmOutlier = F,
  verbose = F, outlierCutOff = 0.995)
}
\arguments{
\item{experiment}{The Hi-C experiment object of a sample: produced by \code{construct.experiment()}.}

\item{tad.bed}{Data.frame from a BED-file containing the TAD positions.}

\item{smallTreshold}{The minimal size of TADs. Too small TADs lead to messy plots.}

\item{rmOutlier}{Perform outlier-correction by tresholding [outlierCutOff] percentile per pixel.}

\item{verbose}{Produces a progress-indication.}

\item{outlierCutOff}{The severity of outliers. We compute the [outlierCutOff] percentile per pixel and set values bigger than that to this value.}
}
\value{
\item{STACK}{a list containing a matrix with the Z-stack scores}

\item{STACK.raw}{a matrix without the outlier-correction}

\item{STACK.list}{raw matrices per TAD}

\item{OUTLIERCORRECTIONSWITCH}{the outlierCutOff if rmOutlier is TRUE, otherwise set to FALSE}
}
\description{
Get a z-stack matrix of TADs: this function extracts matrices from a BED-like structure and resizes them,
which leads to all start- and end-position of TADs overlapping. Next, it averages all matrices to produce
a single Z-stack matrix.
}
\examples{
ATA_results_of_RAOetal <- ATA(experiment = Rao_20k,tad.bed = TADs)
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
