% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badBin.find.R
\name{badBin.find}
\alias{badBin.find}
\title{Find all bad bins}
\usage{
badBin.find(experiment, p.treshold = 0.1, plotBad = F, getBed = T)
}
\arguments{
\item{experiment}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{p.treshold}{Cutoff of spearmans correlation.}

\item{plotBad}{Plot some of the bad bins for visual inspection.}

\item{getBed}{Produce a bed-like object.}
}
\value{
A list containing a vector with the bad bin-id's (`badBins`). Optionally, it can output a bed-file (`bed`).
}
\description{
Builds a representative RCP and tests per Hi-C bin, whether its log-linear fitted data correlates.
}
\examples{
# Identify and plot the bad bins
bb_out <- badBin.find(experiment = WT_100kb, p.treshold = 0.1, getBed = T, plotBad = T)

# plot the region from within hic.matrixplot
hic.matrixplot(exp1 = WT_100kb, chrom = 'chr7',  start = 26.75e6,  end=28.5e6, chip = list(bb_out$bed))
}
