% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bed2idx}
\alias{bed2idx}
\title{Match bed-like entries to Hi-C bin indices}
\usage{
bed2idx(IDX, bed, mode = c("centre", "start", "end"))
}
\arguments{
\item{IDX}{The IDX-slot of a \code{contacts} object}

\item{bed}{A 3-column data.frame}

\item{mode}{A \code{character} of length 1 indicating what position of the
\code{bed} argument to match with the indices. Possible values:
\code{"center"}, \code{"start"} or \code{"end"}.}
}
\value{
An \code{integer} vector of length \code{nrow(bed)} and parallel to
  \code{bed} with indices to the Hi-C matrix.
}
\description{
Match bed-like entries to Hi-C bin indices
}
\details{
Out of bounds values are matched to nearest bin.
}
