% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centromere_telomere.R
\name{centromere.telomere.analysis}
\alias{centromere.telomere.analysis}
\title{centromere.telomere.analysis}
\usage{
centromere.telomere.analysis(exp, chrom.vec, nrow = 100,
  leave.out = NULL, q.top = 1e-05, verbose = F)
}
\arguments{
\item{exp}{GENOVA data structure for your experiment of interest}

\item{chrom.vec}{a vector containing the chromosomes that should be
considered in the pair-wise interchromosomal interactions}

\item{nrow}{dimensions of the matrix (number of columns will be the same)}

\item{leave.out}{two-column matrix or data.frame containing the specific
chromosome combinations that need to left out}

\item{q.top}{top quantile of scores that should be left out of the analysis
(because they are outliers)}

\item{verbose}{Produces a progress-indication.}
}
\description{
Calculate a scaled matrix of the interchromosomal interactions aligned from
centromere to telomere
}
\examples{
# Get a scaled matric of the interchomosomal interactions between 15 and 19
out1519 = centromere.telomere.analysis(WT_40kb,
                                       chrom.vec = c('chr15', 'chr19'))

# Plot the results
draw.centromere.telomere(out1519)
}
\author{
Elzo de Wit, \email{e.d.wit@nki.nl}
}
