% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation_score.R
\name{chromosome.wide.insulation}
\alias{chromosome.wide.insulation}
\title{chromosome.wide.insulation}
\usage{
chromosome.wide.insulation(hic, window.size, chrom)
}
\arguments{
\item{hic}{A GENOVA-experiment}

\item{window.size}{The sliding square size}

\item{chrom}{Chromosome to use}
}
\value{
DF with insulation score
}
\description{
loop over the entire chromosome and calcute the insulation score for every position, use this to correct if requeste
}
