% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomeMatrix.R
\name{chromosomeMatrix}
\alias{chromosomeMatrix}
\title{chromosomeMatrix}
\usage{
chromosomeMatrix(exp, color.fun = NULL, z.max = NULL)
}
\arguments{
\item{color.fun}{Optional color-function}

\item{z.max}{Adjust the maximum value of the color-scale}

\item{experiment}{The Hi-C experiment object of a sample: produced by construct.experiment().}
}
\value{
A plot of the chromosome matrix and a assignable matrix
}
\description{
Plot a genome-wide overview-matrix.
}

