% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_plot.R
\name{cis.compartment.plot}
\alias{cis.compartment.plot}
\title{cis.compartment.plot}
\usage{
cis.compartment.plot(exp1, exp2 = NULL, chrom, arm = "p",
  cut.off = NULL, obs.exp = F, invert = F, color.scheme = "fall",
  cs.lim = NULL, chip = NULL, smoothNA = F)
}
\arguments{
\item{exp1}{A GENOVA experiment-object}

\item{exp2}{A GENOVA experiment-object. If given, all plots show this exp in the lower-left corner.}

\item{chrom}{The chromosome that should be drawn}

\item{arm}{Which chromosome arm: 'p' or 'q' (for acrocentric this should be set to 'q')}

\item{cut.off}{maximum value for the heatmap}

\item{obs.exp}{whether an observed over expected matrix should be drawn (default is false)}

\item{invert}{whether the compartment score should be inverted}

\item{color.scheme}{color scheme that should be used, defaults to fall, other values result in white-red gradient}

\item{cs.lim}{y-axis limit for the compart score, if unset (NULL) will default to the maximum absolute value}

\item{chip}{A data.frame, containg ChIP-seq peaks of active histone marks to correctly orient A/B compartments}

\item{smoothNA}{Set to TRUE to perform a Nadaraya/Watson normalization. This will try to eliminate white stripes: this is only cosmetic and has no effect on the compartment-scores.
Requires fields and is only needed if large white stripes are bothering you.}
}
\description{
Draw intrachromosomal interaction heatmap for a chromosome (arm) with corresponding compartment scores
}
\note{
# Plot a cis-compartment plot of the q-arm of chromosome 14.
cis.compartment.plot(exp = Hap1_WT_40kb, chrom = 'chr14', arm = 'q', cs.lim = 1.75, cut.off = 15, chip = H3K27ac_peaks)

# Plot a observed/expected cis-compartment plot of the q-arm of chromosome 14.
cis.compartment.plot(exp = Hap1_WT_40kb, chrom = 'chr14', arm = 'q', cs.lim = 1.75, cut.off = 15, chip = H3K27ac_peaks, obs.exp = T)
}
