% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cisTotal.scores.R
\name{cisTotal.scores}
\alias{cisTotal.scores}
\title{cisTotal.scores}
\usage{
cisTotal.scores(exp)
}
\arguments{
\item{exp}{The Hi-C experiment object of a sample:
produced by construct.experiment().}
}
\value{
The genome-wide cis/total ratio
}
\description{
Calculate the genome-wide ratio of the intrachromosomal interactions over
the interchromosomal interactions
}
\examples{
# calculate the ratios for a 1Mb-resolution experiment
cis_out <- cisTotal.scores(WT_1MB)
}
\author{
Elzo de Wit, \email{e.d.wit@nki.nl}
}
