% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment.score.R
\name{compartment.score}
\alias{compartment.score}
\title{Get compartment-scores for a chromosome}
\usage{
compartment.score(exp, chrom = "chr2", comparableTrack = NULL)
}
\arguments{
\item{chrom}{The chromosome of interest.}

\item{comparableTrack}{An optional vector, which will be used to flip the A/B signal.
For example, one can use GC-content to ensure that regions with positive scores are postiviely correlated with GC-content.}

\item{experiment}{The Hi-C experiment object of a sample: produced by construct.experiment().}
}
\value{
A vector with the compartment-score for every bin in the Hi-C matrix.
}
\description{
This uses the per-arm compartment-score algorithm of G. Fudenberg.
The procedure is an eigenvector decomposition of the observed/expected matrix.
}
