% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.experiment.R
\name{construct.experiment}
\alias{construct.experiment}
\title{Construct a HiC-experiment.}
\usage{
construct.experiment(signalPath, indicesPath, name, centromeres = NULL,
  color = 1, comments = NULL, BPscaling = 1e+09)
}
\arguments{
\item{signalPath}{Full path to a HiC-pro-like matrix-file or .sig from juicerToGenova.py}

\item{indicesPath}{Full path the HiC-pro-like index-file or .bed from juicerToGenova.py}

\item{name}{The name of the sample.}

\item{centromeres}{A data.frame with three columns per chromosome: chromosome name, start-position and end-position of the centromeric region.}

\item{color}{Color associated with sample.}

\item{BPscaling}{Scale contacts to havebgenome-wide sum of [BPscaling] reads.}

\item{COMMENTS}{A place to store some comments.}
}
\value{
A list.
}
\description{
Make a structure which holds the most needed information of a HiC-experiment.
}
