% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.experiment.R
\name{construct.experiment}
\alias{construct.experiment}
\title{Construct a HiC-experiment.}
\usage{
construct.experiment(signalPath, indicesPath, name, ignore.checks = F,
  centromeres = NULL, color = 1, comments = NULL,
  BPscaling = 1e+09)
}
\arguments{
\item{signalPath}{Full path to a HiC-pro-like matrix-file or .sig from juicerToGenova.py}

\item{indicesPath}{Full path the HiC-pro-like index-file or .bed from juicerToGenova.py}

\item{name}{The name of the sample.}

\item{ignore.checks}{Skip the checks for empty matrices: EXPERT-ONLY!}

\item{centromeres}{A data.frame with three columns per chromosome: chromosome name, start-position and end-position of the centromeric region.}

\item{color}{Color associated with sample.}

\item{comments}{A place to store some comments.}

\item{BPscaling}{Scale contacts to have agenome-wide sum of [BPscaling] reads (default: 1000M). Set to NULL to skip this.}
}
\value{
An experiment-object, which is a named list of contacts, indices and metadata for a Hi-C matrix of a given sample at a given resolution.
}
\description{
Make an object which holds the most needed information of a HiC-experiment at a given resolution.
}
\note{
Some reference genomes have very small "random" or "patch" chromosomes, which can have zero contacts mapped to it (at certain resolutions). Construct.experiment checks this and omits these chromosomes in the resulting experiment-object. The RMCHROM-flag will also be set to TRUE: this will help other GENOVA-functions to deal better with this problem. There is a slight performance-cost during the construction of the experiment, however. Therefore, experienced users can set ignore.checks to TRUE to skip all of this, keeping in mind that some functions will not work properly/at all is there are these zero-coverage chromosomes in their data.
}
\examples{
WT_10kb <- construct.experiment(ignore.checks = T, signalPath = 'WT_10kb_iced.matrix', indicesPath = 'WT_10kb_abs.bed', name = "WT", color = "black")
}
