% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMLA.R
\name{CSCAn}
\alias{CSCAn}
\title{Cross spatial chromatin analysis}
\usage{
CSCAn(
  explist,
  bedlist,
  shift = 1000000L,
  dist_thres = c(NA, 1e+06),
  size_bin = NULL,
  size_bp = NULL,
  outlier_filter = c(0, 1),
  min_compare = 10,
  anchors = NULL,
  raw = TRUE,
  group_direction = FALSE
)
}
\arguments{
\item{explist}{Either a single GENOVA \code{contacts} object or a list of
GENOVA \code{contacts} objects.}

\item{bedlist}{A named \code{list} of length >= 2 wherein each element is a 
BED-like \code{data.frame} with the following three columns: 
\enumerate{ \item A \code{character} giving the chromosome names. \item An
\code{integer} with start positions. \item An \code{integer} with end
positions. }}

\item{shift}{An \code{integer} of length 1 indicating how many basepairs the
anchors should be shifted. Essentially performs circular permutation of
\code{size} for a reasonable estimate of background. The argument is
ignored when \code{shift <= 0}.}

\item{dist_thres}{An \code{integer} vector of length 2 indicating the minimum
and maximum distances in basepairs between anchorpoints.}

\item{size_bin, size_bp}{The size of the lookup regions in bins (i.e. a score
of 21 yields an output with 10 Hi-C bins both up- and downstream of the
anchor). When \code{NULL} (default), it is internally set to \code{21} when
the \code{size_bp} is also \code{NULL}. \code{size_bp} is an alternative
parametrisation for the lookup regions, expressed in basepairs.
\code{size_bp} is not used when the argument \code{size_bin} is set.}

\item{outlier_filter}{A \code{numeric} of length 2 between \code{[0-1]}
indicating quantiles of data to be used as thresholds. Data outside these
thresholds are set to the nearest threshold. Setting this to \code{c(0, 1)}
performs no outlier correction.}

\item{min_compare}{An \code{integer} of length 1 indicating the minimum
number of pairwise interactions on a chromosome to consider.}

\item{anchors}{(Optional) A \code{matrix} with two columns containing
pre-computed anchor indices. If this is set, skips calculation of anchor
indices and uses this argument instead. See
\code{\link[GENOVA:anchors]{anchors_CSCAn()}}.}

\item{raw}{A \code{logical} of length 1: should the raw array underlying the
summary matrices be returned in the output? Should be \code{TRUE} if the
intention is to use the \code{quantify} function.}

\item{group_direction}{A \code{logical} of length 1 which when \code{TRUE}
will mirror groups for anchors, where the left anchor location is larger
than the right anchor location. Left and right refer to the \code{bedlist}
elements that generate combinations.}
}
\value{
An \code{CSCAn_discovery} object containing the following slots, 
  wherein \code{i} is the number of combinations between the \code{bedlist} 
  argument elements:
  \describe{ \item{obsexp}{An \code{array} with the dimensions
  \code{size_bin} x \code{size_bin} x \code{i} x \code{length(explist)} containing fold
  change values for the signal over the median shifted values.}
  \item{signal}{An \code{array} with the dimensions \code{size_bin} x
  \code{size_bin} x \code{i} x \code{length(explist)} containing mean contact values for
  bins surrounding the anchors.} \item{signal_raw}{A \code{list} with
  \code{length(explist)} elements for each contacts object, wherein an
  element is an n x \code{size_bin} x \code{size_bin} array with contact
  values for each anchor. 'n' is the number of non-empty valid anchors.}
  \item{shifted}{An \code{array} with the dimensions \code{size_bin} x
  \code{size_bin} x \code{i} x \code{length(explist)} containing mean contact values for
  bins that are \code{shift} basepairs away from the anchors.}
  \item{shifted_raw}{A \code{list} with \code{length(explist)} elements for
  each contacts object, wherein an alemeent is an n x \code{size_bin} x
  \code{size_bin} array with contact values for each shifted anchors. 'n' is 
  the number of non-empty valid (unshifted) anchors.}}
}
\description{
Takes a list of BED-like genomic locations and makes a Hi-C contact analysis
for crosswise pairs between elements of that list.
}
\examples{
\dontrun{
cscan <- CSCAn(list(WT_20kb, KO_20kb),
               bedlist = list(group_A, group_B))
               
visualise(cscan)
}
}
\seealso{
The \code{\link[GENOVA]{rep_mat_lookup}} function that performs the
  lookup and summary for the \code{CSCan} function and others. \cr The
  \code{\link[GENOVA]{discovery}} class for a general description of
  \code{discovery} classes. \cr The \code{\link[GENOVA]{visualise}} function
  for visualisation of the results. \cr The \code{\link[GENOVA]{anchors}}
  documentation for more information about anchors.

Other aggregate repeated matrix lookup analyses: 
\code{\link{APA}()},
\code{\link{ARA}()},
\code{\link{ATA}()},
\code{\link{PESCAn}()},
\code{\link{rep_mat_lookup}()}
}
\concept{aggregate repeated matrix lookup analyses}
