% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.TAD.scatterplot.R
\name{differential.TAD.scatterplot}
\alias{differential.TAD.scatterplot}
\title{differential.TAD.scatterplot}
\usage{
differential.TAD.scatterplot(exp1, exp2, line = T, allData = T,
  color.fun = NULL, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{exp1}{Output of intra.inter.TAD.contacts for experiment 1.}

\item{exp2}{Output of intra.inter.TAD.contacts for experiment 2.}

\item{line}{Add a diagonal line?}

\item{allData}{Set to FALSE to get a zoomed in view.}

\item{color.fun}{A color-function like rainbow().}

\item{xlab}{The x-axis label.}

\item{ylab}{The y-axis label.}
}
\value{
A scatterplot.
}
\description{
Create a scatter plot of the inter and intra TAD contact frequencies.
}
\examples{
# get scores for WT and WAPL data
TAD_N_WT   <- intra.inter.TAD.contacts(TAD = WT_TADs,
                                       max.neighbor = 10,
                                       exp = WT_10kb)
TAD_N_WAPL <- intra.inter.TAD.contacts(TAD = WT_TADs,
                                       max.neighbor = 10,
                                       exp = WAPL_10kb)

# plot a TAD+N scatterplot
differential.TAD.scatterplot(exp1 = TAD_N_WT, # x
                             exp2 = TAD_N_WAPL, # y
                             allData = T)
}
\author{
Elzo de Wit, \email{e.d.wit@nki.nl}
}
