% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centromere_telomere.R
\name{draw.centromere.telomere}
\alias{draw.centromere.telomere}
\title{draw.centromere.telomere}
\usage{
draw.centromere.telomere(m, cut.off = 2)
}
\arguments{
\item{m}{matrix result from centromere-telomere}

\item{cut.off}{log2-ratio cut.off}
}
\description{
Plot the log2-ratio of the relative interaction frequencies.
The ratio is calculated over the median of the matrix.
}
\examples{
# Get a scaled matric of the interchomosomal interactions between 15 and 19
out1519 = centromere.telomere.analysis(WT_40kb,
                                       chrom.vec = c('chr15', 'chr19'))

# Plot the results
draw.centromere.telomere(out1519)
}
\author{
Elzo de Wit, \email{e.d.wit@nki.nl}
}
