% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_matrix_lookup.R
\name{engine_resizing}
\alias{engine_resizing}
\title{Engine for repeated matrix lookup and resizing}
\usage{
engine_resizing(MAT, anchors, rel_pos)
}
\arguments{
\item{anchors}{A \code{matrix} with two columns containing pre-computed
indices between which a square is looked up.}

\item{rel_pos}{A \code{integer} sequence ranging from \code{[0-n]} wherein
  \code{n} is the target size to which individual squares are resized.

  \code{anchors} specify two points on the diagonal between which the region
  should be looked up.}

\item{ICE}{The Hi-C matrix slot of a GENOVA experiment object.}
}
\value{
A threedimensional \code{array} wherein the first dimension is
  parallel to the rows in \code{anchors}, and the second and third dimensions
  are of length \code{m}.
}
\description{
It looks up square parts of unequal size around the diagonal of a single Hi-C
experiment and resizes these to a single size.
}
\seealso{
\code{\link[GENOVA]{rep_mat_lookup}}
}
\keyword{internal}
