% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.interp.surface.R
\name{fields.interp.surface}
\alias{fields.interp.surface}
\title{fields.interp.surface}
\usage{
fields.interp.surface(obj, loc)
}
\arguments{
\item{obj}{A surface or image  object like the list for contour, persp or image.}

\item{lof}{A matrix of 2 d locations -- new points to evaluate the surface.}
}
\value{
An interpolated matrix
}
\description{
fields.interp.surface
}
\note{
THIS FUNCTION IS TAKEN FROM FIELDS!

fields is a package for analysis of spatial data written for
the R software environment .
Copyright (C) 2016
University Corporation for Atmospheric Research (UCAR)
Contact: Douglas Nychka, nychka@ucar.edu,
National Center for Atmospheric Research, PO Box 3000, Boulder, CO 80307-3000

This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 2 of the License, or
(at your option) any later version.
This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You should have received a copy of the GNU General Public License
along with the R software environment if not, write to the Free Software
Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
or see http://www.r-project.org/Licenses/GPL-2
}
