% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiC_matrixplot.R
\name{hic.matrixplot}
\alias{hic.matrixplot}
\title{hic.matrixplot}
\usage{
hic.matrixplot(exp1, exp2 = NULL, chrom, start, end, cut.off = 0,
  chip = list(NULL, NULL, NULL, NULL), inferno = T, bed.col = rep(c("red",
  "blue"), 2), bw.col = rep(c("navy", "darkred"), 2), type = rep("triangle",
  4), coplot = "dual", genes = NULL, tads = NULL, tad.type = "lower",
  loops = NULL, loops.type = "lower", tads.color = "#3288bd",
  loops.resize = 0, loops.color = "#3288bd")
}
\arguments{
\item{exp1}{The control Hi-C experiment object: produced by construct.experiment(). (bottom)}

\item{exp2}{Optional: the treatment Hi-C experiment object: produced by construct.experiment().}

\item{chrom}{Chromosome}

\item{start}{Start position of the region of interest}

\item{end}{End position of the region of interest}

\item{cut.off}{The cut.off for the hic-matrix plot, in the diff option the negative of this is the lower bound}

\item{chip}{A list of feature tracks, can be bed structure (i.e. data frames) or a path to bigwig file (i.e. character variable), maximum length of 4, first two and last two have to be the same (i.e. 1:bed,2:bed,3:path,4:path, but 1:bed,2:NULL,3:path,4:NULL is also allowed)}

\item{inferno}{White/Red/black or White/Red coloscale?}

\item{bed.col}{Color of the bed track (max.len is 4)}

\item{bw.col}{Same as bed col, but for the bigwig track}

\item{type}{Should a rectangle or a triangle be drawn? Not that for a triangle a 6th strand column should be included}

\item{coplot}{When drawing together two experiments, dual is bottom triangle exp1, top triangle exp2; diff plots a substraction of exp2-exp1}

\item{genes}{Structure with gene information, will only be combined with bed structure}

\item{tads}{BED-like dataframe (not a tibble or a data.table!)}

\item{tad.type}{How to show TADS: upper, lower and or both}

\item{loops}{BED-like dataframe (not a tibble or a data.table!)}

\item{loops.type}{How to show loops: upper, lower and or both}

\item{loops.resize}{Make the loop-boxes bigger by X bp. Can help visibility.}

\item{loops.color}{Which color do you want the loops to have?}
}
\value{
A matrix-plot
}
\description{
Plot a matrix (or two) for a region of interest with annotations
}
