% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiC_matrixplot.R
\name{hic.matrixplot}
\alias{hic.matrixplot}
\title{hic.matrixplot}
\usage{
hic.matrixplot(exp1, exp2 = NULL, chrom, start, end, cut.off = NULL,
  chip = list(NULL, NULL, NULL, NULL), inferno = T, cexTicks = 1,
  bed.col = rep("black", 4), bw.col = "black", yMax = NULL,
  type = rep("triangle", 4), guessType = T, coplot = "dual",
  genes = NULL, tads = NULL, tads.type = "lower", loops = NULL,
  loops.type = "lower", tads.color = "#3288bd", loops.resize = 0,
  loops.color = "#3288bd", skipAnn = F, symmAnn = F, check.genome = T,
  antoni = F)
}
\arguments{
\item{exp1}{The control Hi-C experiment object: produced by construct.experiment(). (bottom)}

\item{exp2}{Optional: the treatment Hi-C experiment object: produced by construct.experiment().}

\item{chrom}{Chromosome}

\item{start}{Start position of the region of interest}

\item{end}{End position of the region of interest}

\item{cut.off}{The cut.off for the hic-matrix plot, in the diff option the negative of this is the lower bound}

\item{chip}{A list of feature tracks, can be bed structure (i.e. data frames) or a path to bigwig file (i.e. character variable), maximum length of 4. Placement is inner-top, outer-top, outer-left, inner-left.}

\item{inferno}{White/Red/black or White/Red coloscale?}

\item{cexTicks}{Change size of numbers on the axis}

\item{bed.col}{Color of the bed track (max.len is 4)}

\item{bw.col}{Same as bed col, but for the bigwig track}

\item{yMax}{Set the maximum height of all biwigs.}

\item{type}{Should a rectangle or a triangle be drawn? Note that for a triangle a 6th strand column should be included}

\item{guessType}{If an element in the chip is a dataframe, infer the type from column 6? If true, column six should hold "+" and "-". If a row has other characters, we view this entry as no-oriention and thus plot a rectangle.}

\item{coplot}{When drawing together two experiments, dual is bottom triangle exp1, top triangle exp2; diff plots a substraction of exp2-exp1}

\item{genes}{Structure with gene information, will only be combined with bed structure}

\item{tads}{BED-like dataframe or a list of these data.frames (not a tibble or a data.table!)}

\item{tads.type}{How to show TADS: upper, lower and or both}

\item{loops}{BED-like dataframe or a list of these data.frames (not a tibble or a data.table!)}

\item{loops.type}{How to show loops: upper, lower and or both}

\item{loops.resize}{Make the loop-boxes bigger by X bp. Can help visibility.}

\item{loops.color}{Which color do you want the loops to have?}

\item{skipAnn}{Do not plot outside-annotation. Can be used to plot other things next to the matrix.}

\item{symmAnn}{Put features 1&2 in also on verical (ignore chip-entries 3 & 4)}

\item{check.genome}{Check if reference genomes in exp1 and exp2 are the same}

\item{antoni}{Logical: plot an explorer of the microscopic world}
}
\value{
A matrix-plot
}
\description{
Plot a matrix (or two) for a region of interest with annotations
}
