% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation.callTAD.R
\name{insulation.callTAD}
\alias{insulation.callTAD}
\title{insulation.callTAD}
\usage{
insulation.callTAD(exp, BEDCOLOR = "127,201,127")
}
\arguments{
\item{BEDCOLOR}{Color of items in the resulting BEDPE-file}

\item{exp1}{The Hi-C experiment object: produced by construct.experiment().}
}
\value{
A list, with delta-values, border-calls and a BEDPE-file
}
\description{
Call TADs
}
\note{
Call insulation scores first and store these in exp$INSULATION
}

