% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation_score.R
\name{insulation.domainogram}
\alias{insulation.domainogram}
\title{insulation.domainogram}
\usage{
insulation.domainogram(exp1, chrom, start, end, window.size1 = 5,
  window.size2 = 101, step = 2, local = T, zlim = c(-1, 0.5))
}
\arguments{
\item{exp1}{A Hi-C experiment object: produced by construct.experiment().}

\item{chrom}{Chromosome}

\item{start}{Start position of the region of interest}

\item{end}{End position of the region of interest}

\item{window.size1}{The minimal sliding square size}

\item{window.size2}{The maximal sliding square size}

\item{step}{Thesliding square step-size}

\item{local}{Per chromosome?}

\item{zlim}{Zlims of the plot}

\item{cut.off}{The cut.off for Hi-C scores}
}
\value{
A plot
}
\description{
create a domainogram for the insulation scores
}

