% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulation_score.R
\name{insulation.score}
\alias{insulation.score}
\title{insulation.score}
\usage{
insulation.score(hic, window.size, chrom, start, end, diag.add = 0,
  local = T)
}
\arguments{
\item{window.size}{The sliding square size}

\item{chrom}{Chromosome}

\item{start}{Start position of the region of interest}

\item{end}{End position of the region of interest}

\item{diag.add}{Add values to diaginal}

\item{local}{Per chromosome?}

\item{exp1}{A Hi-C experiment object: produced by construct.experiment().}

\item{cut.off}{The cut.off for Hi-C scores}

\item{step}{Thesliding square step-size}
}
\value{
A plot
}
\description{
Get insulation scores
}

