% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PESCAn.R
\name{PESCAn}
\alias{PESCAn}
\title{PESCAn}
\usage{
PESCAn(experiment, bed, minDist = 5e+06, size = 5e+05, add = 0)
}
\arguments{
\item{experiment}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{bed}{A bed-dataframe.}

\item{minDist}{The minimal distance}

\item{size}{The amount of Hi-C bins to take into account (i.e. a score of 21 yield an output with 10 Hi-C bins up- and downstream of the anchor).}

\item{add}{Add constant value to bed-start and -end.}
}
\value{
A score-matrix.
}
\description{
From a ChIP bed file, a HiCpro matrix and a HiCpro bed file calculate a PE-scan like data structure. Run this per chromosome.
}

