% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PESCAn_APA.R
\name{PESCAn}
\alias{PESCAn}
\title{Paired-end spatial chromatin analysis}
\usage{
PESCAn(explist, bed, shift = 1000000L, dist_thres = c(5000000L, Inf),
  size_bin = NULL, size_bp = 4e+05, outlier_filter = c(0, 1),
  min_compare = 10, anchors = NULL, raw = FALSE)
}
\arguments{
\item{bed}{A \code{data.frame} with 3 columns in BED format, containing the
regions to anchor in pairwise manner.}

\item{shift}{An \code{integer} of length 1 indicating how many basepairs the
anchors should be shifted. Essentially performs circular permutation of
\code{size} for a reasonable estimate of background. The argument is
ignored when \code{shift <= 0}.}

\item{min_compare}{An \code{integer} of length 1 indicating the minimum
number of pairwise interactions on a chromosome to consider.}
}
\value{
A \code{list} of the same length as \code{explist} wherein list
  elements contain the results of the PE-SCAn per experiment.
}
\description{
Performs an all-to-all Hi-C contact analysis for specified regions.
}
\examples{
# Typical usage: PESCAn for super enhancers using a 1 MB
# circular permutation on a pair of experiments.
pescan <- PESCAn2(explist = list(WT_40kb, KO_40kb),
                  bed = super_enhancers,
                  shift = 1e6)

# Alternative usage with pre-calculated anchors and no permutation
anchors <- anchors_PESCAn(WT_40kb$ABS, WT_40kb$RES,
                          genes_tss,
                          dist_thres = c(5e6, 15e6))
pescan <- PESCAn2(explist = list(WT_40kb),
                  anchors = anchors,
                  shift = 0)
}
