% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PESCAn.R
\name{PESCAn}
\alias{PESCAn}
\title{PE-SCAn}
\usage{
PESCAn(exp, bed, shift = 1e+06, minComparables = 10, mindist = 5e+06,
  maxDist = Inf, size = 4e+05, rmOutlier = F,
  outlierCutOff = 0.995, verbose = T)
}
\arguments{
\item{exp}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{bed}{A bed-dataframe.}

\item{shift}{Set to X bp for circular permutation. Set to zero for just getting the signal-matrix.}

\item{minComparables}{Minimum amount of BED-entries per chromosome.}

\item{maxDist}{The maximal distance}

\item{size}{Size in bp of window.}

\item{rmOutlier}{Perform outlier-correction}

\item{outlierCutOff}{The severity of outliers. We compute the [outlierCutOff]
percentile per pixel and set values bigger than that to this value.}

\item{minDist}{The minimal distance}
}
\value{
An list with a O/E score-matrix (if shift != 0), otherwise an observed
score-matrix, the underlying signal and background-matrices and the shift used.
}
\description{
From a ChIP-peaks.BED dataframe, calculate the all-vs-all Hi-C contacts.
}
\examples{
# Run PE-SCAn on a bed of super-enhancers,
using WT Hi-C data and a circular permutation of 1Mb
WT_PE_OUT = PESCAn(exp = WT_40kb,
                   bed = superEnhancers,
                   shift = 1e6)

# Plot using visualise.PESCAn.ggplot
visualise.PESCAn.ggplot(PESCAnlist = list(WT = WT_PE_OUT),
                        resolution = 40e3,
                        smooth = F)

# Plot using persp
RES = 40e3 # resolution of the Hi-C
persp(list(x = seq(-1*(RES*10),(RES*10), length.out = 21)/1e6, # x-ticks (MB)
           y = seq(-1*(RES*10),(RES*10), length.out = 21)/1e6, # y-ticks (MB)
           z = WT_PE_OUT)
}
\author{
Elzo de Wit, \email{e.d.wit@nki.nl}
}
