% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PESCAn.R
\name{PESCAn_covert}
\alias{PESCAn_covert}
\title{PESCAn_covert}
\usage{
PESCAn_covert(experiment, bed, minComparables = 10, rmOutlier = F,
  minDist = 5e+06, maxDist = Inf, size = 5e+05, add = 0,
  outlierCutOff = 0.995, verbose = T)
}
\arguments{
\item{experiment}{The Hi-C experiment object of a sample: produced by construct.experiment().}

\item{bed}{A bed-dataframe.}

\item{minComparables}{The minimal amount of bed-entries for a given chromosome. If this threshold is not reached, PE-SCAn will skip this chromosome.}

\item{rmOutlier}{to outlier-correction}

\item{minDist}{The minimal distance}

\item{maxDist}{The maximal distance}

\item{size}{The amount of Hi-C bins to take into account (i.e. a score of 21 yield an output with 10 Hi-C bins up- and downstream of the anchor).}

\item{add}{Add constant value to bed-start and -end.}

\item{outlierCutOff}{outlierCutOff}
}
\value{
A score-matrix.
}
\description{
From a ChIP bed file, a HiCpro matrix and a HiCpro bed file calculate a PE-scan like data structure. Run this per chromosome.
}
\examples{
# Run PE-SCAn on a bed of super-enhancers, using WT Hi-C data.
WT_PE_OUT = PESCAn(exp = WT_40kb, bed = superEnhancers)

# Plot using visualise.PESCAn.ggplot
visualise.PESCAn.ggplot(PESCAnlist = list(WT = WT_PE_OUT),
                        resolution = 40e3,
                        smooth = F)

# Plot using persp
persp(SE_vs_WT/SE_vs_WT_perm, phi = 30, theta = 30, col = 'skyblue')

}
