% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_quantify.R
\name{quantify}
\alias{quantify}
\alias{quantify.default}
\alias{quantify.APA_discovery}
\alias{quantify.saddle_discovery}
\alias{quantify.IIT_discovery}
\title{Quantification of results}
\usage{
quantify(discovery, ...)

\method{quantify}{APA_discovery}(discovery, signal_size = 3, ...)

\method{quantify}{saddle_discovery}(discovery, ...)

\method{quantify}{IIT_discovery}(discovery, ...)
}
\arguments{
\item{discovery}{A \code{discovery} object as returned by GENOVA analysis functions.}

\item{...}{further arguments passed to or from other methods.
take the middle 3x3 matrix of the APA).}

\item{signal_size}{The width/height of the signal (e.g. a value of 3 will}
}
\description{
Here be a placholder description
}
\examples{
NULL
}
