% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_quantify.R
\name{quantify}
\alias{quantify}
\alias{quantify.default}
\title{Quantification of results}
\arguments{
\item{discovery}{A \code{discovery} object as returned by GENOVA analysis functions.}

\item{...}{further arguments passed to or from other methods.
take the middle 3x3 matrix of the APA).}

\item{signal_size}{The width/height of the signal (e.g. a value of 3 will}
}
\description{
Here be a placholder description
}
