% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCP.R
\name{RCP}
\alias{RCP}
\title{Get a per-chromosome scaling plot}
\usage{
RCP(experimentList, chromsToUse = NULL, maxDistance = 5e+08, verbose = F,
  scaling = F)
}
\arguments{
\item{chromsToUse}{A vector containing the chromosome-names of interest.}

\item{maxDistance}{The maximal distance to calculate scalings for.}

\item{verbose}{Produces a progress-indication.}

\item{scaling}{Generate scaling instead of RCP?}

\item{experiment}{List of experiment-objects from `construct.experiment()`.}
}
\value{
A data_frame with distance-bin and probabilities.
}
\description{
Produce a dataframe with the probabilities of contacts in a set of distance-bins.
Bins are created on a log scale, which leads to equal amounts of datapoints per bin.
}
