% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCP2.R
\name{RCP}
\alias{RCP}
\title{Get a per-chromosome scaling plot}
\usage{
RCP(experimentList, chromsToUse = NULL, bedList = NULL,
  maxDistance = NULL, ignoreLengthWarning = F, outlierCutoff = 1,
  verbose = F)
}
\arguments{
\item{chromsToUse}{A vector containing the chromosome-names of interest. If bed is also given, only entries in chromsToUse will be used.}

\item{bedList}{A named list of BED-like dataframse of the regions of interest. RCP will intersect this with the Hi-C bin by intersecting the middle of both ranges.}

\item{maxDistance}{The maximal distance to calculate scalings for.}

\item{ignoreLengthWarning}{Force RCP to use maxDistance instead of longest chromosome if that is smaller than maxDistance.}

\item{outlierCutoff}{Percentage for the cutoff.}

\item{verbose}{Produces a progress-indication.}

\item{experiment}{List of experiment-objects from `construct.experiment()`.}
}
\value{
A data_frame with distance-bin and probabilities.
}
\description{
Produce a dataframe with the probabilities of contacts in a set of distance-bins.
Bins are created on a log scale, which leads to equal amounts of datapoints per bin.
}
