% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCP.R
\name{RCP}
\alias{RCP}
\title{Relative Contact Probabilities}
\usage{
RCP(
  explist,
  bedlist = NULL,
  chromsToUse = NULL,
  maxDistance = NULL,
  genomeWide = NULL
)
}
\arguments{
\item{explist}{List of  GENOVA contacts-objects from `loadContacts()`.}

\item{bedlist}{A named list of BED-like dataframes of the regions of interest.}

\item{chromsToUse}{A vector containing the chromosome-names of interest.}

\item{maxDistance}{The maximal distance to calculate RCPs for.}

\item{genomeWide}{Normalise with genome-wide counts or per-chromosome (default).}
}
\value{
An RCP_discovery object containing:

\item{raw}{a per-distance probability}

\item{smooth}{a log10-mean smoothed probability}
}
\description{
Produce a dataframe with the probabilities of contacts in a set of distance-bins.
Bins are created on a log scale, which leads to equal amounts of datapoints per bin.
}
\section{Resolution recommendation}{
 40kb-500kb
}

\examples{
# Calculate the RCP of chromosome 1
\dontrun{
RCP_out = RCP(experimentList = list('WT' = WT_1MB), chromsToUse = 'chr1')

# Plot the RCP
visualise(RCP_out)
}
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
