% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCP2.R
\name{RCP}
\alias{RCP}
\title{Relative Contact Probabilities}
\usage{
RCP(experimentList, chromsToUse = NULL, bedList = NULL,
  colors = NULL, maxDistance = NULL, ignoreLengthWarning = F,
  outlierCutoff = 1, verbose = F)
}
\arguments{
\item{chromsToUse}{A vector containing the chromosome-names of interest. If bedList is also given, only BED-entries in chromsToUse will be used.}

\item{bedList}{A named list of BED-like dataframse of the regions of interest. RCP will intersect this with the Hi-C bin by intersecting the middle of both ranges.}

\item{colors}{Override colors of experiments. Will be used for BEDs if only one experiment is given.}

\item{maxDistance}{The maximal distance to calculate RCPs for.}

\item{ignoreLengthWarning}{Force RCP to use maxDistance instead of longest chromosome if that is smaller than maxDistance.}

\item{outlierCutoff}{Remove the [outlierCutoff] top and bottom percentages.}

\item{verbose}{Produces a progress-indication.}

\item{experiment}{List of experiment-objects from `construct.experiment()`.}
}
\value{
A data.frame, containing:

\item{distance}{the distance-bin}

\item{prob}{the relative contact probability}

\item{SEM}{the SEM of the RCP}

\item{BED}{if and which BED is used}

\item{color}{the corresponding color of the lines, taken from exp$COL as default}

\item{sample}{the corresponding sample-name}

\item{chrom}{the corresponding}
}
\description{
Produce a dataframe with the probabilities of contacts in a set of distance-bins.
Bins are created on a log scale, which leads to equal amounts of datapoints per bin.
}
\examples{
# Calculate the RCP of chromosome 1
RCP_out = RCP(experimentList = list('WT' = WT_1MB),
chromsToUse = 'chr1')

# Plot the RCP
visualise.RCP.ggplot(RCP_out)
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
