% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_matrix_lookup.R
\name{rep_mat_lookup}
\alias{rep_mat_lookup}
\title{Repeated Hi-C matrix lookup}
\usage{
rep_mat_lookup(
  explist,
  anchors,
  rel_pos,
  shift = 0,
  outlier_filter = c(0, 1),
  raw = FALSE
)
}
\arguments{
\item{explist}{Either a single GENOVA \code{contacts} object or a list of
GENOVA \code{contacts} objects.}

\item{anchors}{A \code{matrix} with two columns containing pre-computed
anchor indices. See the \code{\link[GENOVA]{anchors}} documentation.}

\item{rel_pos}{An \code{integer} vector indicating relative positions in
bins, indicating ranges around anchors to lookup.}

\item{shift}{An \code{integer} of length 1 indicating how many basepairs the
anchors should be shifted. Essentially performs circular permutation of
\code{size} for a reasonable estimate of background. The argument is
ignored when \code{shift <= 0}.}

\item{outlier_filter}{A \code{numeric} of length 2 between \code{[0-1]}
indicating quantiles of data to be used as thresholds. Data outside these
thresholds are set to the nearest threshold. Setting this to \code{c(0, 1)}
performs no outlier correction.}

\item{raw}{A \code{logical} of length 1: should the raw array underlying the
summary matrices be returned in the output? Should be \code{TRUE} if the
intention is to use the \code{quantify} function.}
}
\value{
A \code{list} of the same length as \code{explist} wherein list
  elements contain the results of the repeated lookup per experiment.
}
\description{
Analyses like aggregrate peak analysis (APA) and paired-end spatial chromatin
analysis (PE-SCAn) require to look up regions of a Hi-C matrix repeatedly.
This function acts as a common method for the lookup and summary of such
repeated lookups for these analyses.
}
\details{
For each row in the \code{anchors} argument a region of the Hi-C
  matrix is looked up corresponding to that anchor. This data is then
  summarised by taking the mean of each position relative to the anchor
  across all the anchors.

  Anchors are subject to a filtering step wherein anchors are discarded when
  they are within \code{rel_pos} range of a chromosome start or end. This
  ensures the anchors all report data from the same chromosome in the x- or
  y-direction.

  For shifted anchors, an attempt is made to shift the anchors in the
  opposite direction before they are discarded.

  When a region corresponding to a non-shifted anchor is looked up and is
  found to have no contacts within that region, it is discarded. Shifted
  regions are only discarded when the corresponding non-shifted anchor is
  discarded.

  Anchors typically contain a '\code{type}' attribute which informs
  \code{rep_mat_lookup} how the lookup should occur. The \code{APA},
  \code{PESCAn} and \code{ARA} anchors look up regions of dimensions
  \code{length(rel_pos)} x \code{length(rel_pos)}. The \code{ARA} anchors
  transpose these square regions when given a '\code{-}' direction before
  summary occurs. The \code{ATA} anchors look up \code{anchors[, 2] -
  anchors[, 1]} sized square regions and resizes these to a
  \code{max(rel_pos)} square region through bilinear interpolation before
  summary.
}
\section{Resolution recommendation}{
 10kb-40kb
}

\seealso{
Other aggregate repeated matrix lookup analyses: 
\code{\link{APA}()},
\code{\link{ARA}()},
\code{\link{ATA}()},
\code{\link{CSCAn}()},
\code{\link{PESCAn}()}
}
\concept{aggregate repeated matrix lookup analyses}
