% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffleHiC.R
\name{shuffleHiC}
\alias{shuffleHiC}
\title{Shuffle a symmetric matrix}
\usage{
shuffleHiC(MAT, symmetric = T)
}
\arguments{
\item{MAT}{A Hi-C matrix: can be produced by `select.subset()`.}

\item{symmetric}{In some special cases you might not want to get a symmetric matrix back. Set this argument on FALSE.}
}
\value{
A shuffled version of the input: either a list with the shuffled `z` (when using `select.subset`) or a shuffled matrix.
}
\description{
ShuffleHiC shuffles a matrix from `select.subset()`` diagonal-wise. That is, every diagonal in the matrix is shuffled independently.
}
\examples{
## Shuffle a region on chromosome 1
HiCregion <- select.subset(EXP, chrom = 'chr1', start = 15e6, end = 20e6)
HiCregion_shuffled <- shuffleHiC(HiCregion)

# plot observed/expected
image(HiCregion$z/HiCregion_shuffled$z)
}
