% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_bundeling.R
\name{unbundle}
\alias{unbundle}
\alias{unbundle.ARMLA_discovery}
\alias{unbundle.domainogram_discovery}
\alias{unbundle.IS_discovery}
\alias{unbundle.virtual4C_discovery}
\alias{unbundle.CS_discovery}
\alias{unbundle.saddle_discovery}
\alias{unbundle.RCP_discovery}
\alias{unbundle.DI_discovery}
\alias{unbundle.IIT_discovery}
\alias{unbundle.chrommat_discovery}
\title{Split discovery objects}
\usage{
unbundle(discovery, ...)

\method{unbundle}{ARMLA_discovery}(discovery, ...)

\method{unbundle}{domainogram_discovery}(discovery, ...)

\method{unbundle}{IS_discovery}(discovery, ...)

\method{unbundle}{virtual4C_discovery}(discovery, ...)

\method{unbundle}{CS_discovery}(discovery, ...)

\method{unbundle}{saddle_discovery}(discovery, ...)

\method{unbundle}{RCP_discovery}(discovery, ...)

\method{unbundle}{DI_discovery}(discovery, ...)

\method{unbundle}{IIT_discovery}(discovery, ...)

\method{unbundle}{chrommat_discovery}(discovery, ...)
}
\arguments{
\item{discovery}{A \code{discovery} object with more than 1 sample.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A \code{list} wherein each element is a \code{discovery} object for a
  single sample.
}
\description{
\code{unbundle} takes a \code{discovery} object and splits these
  out to individual experiments.
}
\details{
In case the \code{discovery} contains incomplete samples with
  missing slots, \code{NULL} is returned.
}
\examples{
\dontrun{
# Getting a discovery object
apa <- APA(list(WT_20kb, KO_20kb), loops)

# Splitting the results
split <- unbundle(apa)

# Plotting the first result only
visualise(split[[1]])
}
}
\seealso{
\code{\link[GENOVA]{bundle}} for the opposite of \code{unbundle}.
  The \code{\link[GENOVA]{discovery}} class.
}
