% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_4C.R
\name{virtual_4C}
\alias{virtual_4C}
\title{Get an average virtual 4C of 2D-regions (e.g. loops).}
\usage{
virtual_4C(explist, viewpoint, xlim = NULL)
}
\arguments{
\item{explist}{Either a single GENOVA \code{contacts} object or list of
GENOVA \code{contacts}.}

\item{viewpoint}{A viewpoint location. One of the following: \itemize{
  \item{A 3-column, 1-row \code{data.frame} in BED-format.}
  \item{A single \code{character} describing a locus in UCSC notation, e.g. 
  \code{"chr1:30,000,000-40,000,000"}.}
}}

\item{xlim}{A vector of two with the flanking basepairs up- and downstream of
the viewpoint, resp.}
}
\value{
A virtual4C_discovery object.
}
\description{
Extract matrices around a defined region a.k.a. the viewpoint
}
