% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.APA.ggplot.R
\name{visualise.APA.ggplot}
\alias{visualise.APA.ggplot}
\title{Plot the APA-results}
\usage{
visualise.APA.ggplot(APAlist, resolution, title = "APA", zTop = NULL,
  zBottom = NULL, focus = 1, ...)
}
\arguments{
\item{APAlist}{A list of results from `APA`.}

\item{title}{Text to plot}

\item{zTop}{The min and max colorscale-values for the first row of plots.}

\item{zBottom}{The min and max colorscale-values for the first row of plots.}

\item{Focus}{Which sample does need to be the to-compare sample?}
}
\value{
A grid object, containing two ggplot-objects.
}
\description{
Plot the APA-results
}
