% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.APA.ggplot.R
\name{visualise.APA.ggplot}
\alias{visualise.APA.ggplot}
\title{visualise.APA.ggplot}
\usage{
visualise.APA.ggplot(APAlist, title = "APA", zTop = NULL,
  zBottom = NULL, focus = 1, ...)
}
\arguments{
\item{APAlist}{A list of results from the APA-function.}

\item{title}{Title-text to plot.}

\item{zTop}{The min and max values for the first row of plots.}

\item{zBottom}{The min and max values for the first row of plots.}

\item{focus}{Which sample will be the to-compare sample in the differential row?}
}
\value{
A grid object, containing two ggplot-objects.
}
\description{
Plot the APA-results and the differential results.
}
\details{
By substracting the values from each sample with the values from the \code{focus}-sample, we generate the differentials.
A positive value in the differential plots thus means an enrichment in that sample versus the \code{focus}-sample.
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
