% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddle.R
\name{visualise.compartmentStrength}
\alias{visualise.compartmentStrength}
\title{Plot compartment-strengths}
\usage{
visualise.compartmentStrength(SBoutList)
}
\arguments{
\item{SBoutList}{A list of outputs trom saddleBins.}
}
\value{
A plot and an invisible underlying dataframe containing per-sample and -chromosome-arm the compartment-score.
}
\description{
Takes the input from saddleBins and produces a boxplot of compartment-strenghts.
}
\examples{
# run saddleBins on all chromosomes with 25 bins.
saddle_WT = saddleBins(exp = Hap1_WT_1MB, ChIP = H3K27acPeaks,nBins = 25)

# plot compartment-strengths
CS = visualise.compartmentStrength(list(saddle_WT))
median(tmp$score)

}
