% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddleFunctions.R
\name{visualise.compartmentStrength}
\alias{visualise.compartmentStrength}
\title{Plot compartment-strengths}
\usage{
visualise.compartmentStrength(SBoutList, showInteractions = F)
}
\arguments{
\item{SBoutList}{A list of outputs trom saddle().}

\item{showInteractions}{Instead of compartment-strength, plot the average AA, BB and AB interactions.}
}
\value{
A plot and an invisible underlying dataframe containing per-sample and -chromosome-arm the compartment-score.
}
\description{
Takes the input from saddle() and produces a boxplot of compartment-strenghts.
}
\examples{
# run saddle() on all chromosomes with 25 bins.
saddle_WT = saddle(exp = Hap1_WT_1MB, chip = H3K27acPeaks,nBins = 25)

# plot compartment-strengths
CS_out = visualise.compartmentStrength(list(saddle_WT))

}
