% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PESCAn.R
\name{visualise.PESCAn.ggplot}
\alias{visualise.PESCAn.ggplot}
\title{visualise.PESCAn.ggplot}
\usage{
visualise.PESCAn.ggplot(PESCAnlist, resolution, title = "PE-SCAn",
  zTop = NULL, zBottom = NULL, focus = 1, smooth = F, ...)
}
\arguments{
\item{PESCAnlist}{A list of results from PESCAn. Should have the same [shift].}

\item{title}{Text to plot}

\item{zTop}{The min and max values for the first (observed or OE) row of plots.}

\item{zBottom}{The min and max values for the bottom (differential) row of plots.}

\item{Focus}{Which sample does need to be the to-compare sample?}
}
\value{
A grid object, containing two ggplot-objects.
}
\description{
Plot the PE-SCAn-results and differentials.
}
\examples{
# Run PE-SCAn on a bed of super-enhancers,
using WT Hi-C data and a circular permutation of 1Mb
WT_PE_OUT = PESCAn(exp = WT_40kb,
                   bed = superEnhancers,
                   shift = 1e6)

# Plot using visualise.PESCAn.ggplot
visualise.PESCAn.ggplot(PESCAnlist = list(WT = WT_PE_OUT),
                        resolution = 40e3,
                        smooth = F)
}
\author{
Robin H. van der Weide, \email{r.vd.weide@nki.nl}
}
