% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.RCP.ggplot.R
\name{visualise.RCP.ggplot}
\alias{visualise.RCP.ggplot}
\title{Plot the RCP results}
\usage{
visualise.RCP.ggplot(RCPdata, smooth = F, combine = T)
}
\arguments{
\item{RCPdata}{Output of `RCP`}

\item{smooth}{Plot a lowess-smoothed line?}

\item{combine}{Combine all chromosomes?}
}
\value{
A ggplot-object
}
\description{
Produces a line-plot per chromosome.
}

