% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods_visualise.R
\name{visualise}
\alias{visualise}
\alias{visualise.default}
\alias{visualise.APA_discovery}
\alias{visualise.CSCAn_discovery}
\alias{visualise.PESCAn_discovery}
\alias{visualise.ATA_discovery}
\alias{visualise.ARA_discovery}
\alias{visualise.CS_discovery}
\alias{visualise.IS_discovery}
\alias{visualise.DI_discovery}
\alias{visualise.RCP_discovery}
\alias{visualise.virtual4C_discovery}
\alias{visualise.saddle_discovery}
\alias{visualise.domainogram_discovery}
\alias{visualise.IIT_discovery}
\title{Visualise discoveries}
\usage{
visualise(discovery, ...)

\method{visualise}{APA_discovery}(
  discovery,
  contrast = 1,
  metric = c("diff", "lfc"),
  raw = FALSE,
  title = NULL,
  colour_lim = NULL,
  colour_lim_contrast = NULL,
  show_single_contrast = FALSE,
  ...
)

\method{visualise}{CSCAn_discovery}(
  discovery,
  mode = c("obsexp", "signal"),
  raw = FALSE,
  title = NULL,
  colour_lim = NULL,
  show_single_contrast = FALSE,
  ...
)

\method{visualise}{PESCAn_discovery}(
  discovery,
  contrast = 1,
  metric = c("diff", "lfc"),
  mode = c("obsexp", "signal"),
  raw = FALSE,
  title = NULL,
  colour_lim = NULL,
  colour_lim_contrast = NULL,
  show_single_contrast = FALSE,
  ...
)

\method{visualise}{ATA_discovery}(
  discovery,
  contrast = 1,
  metric = c("diff", "lfc"),
  raw = FALSE,
  title = NULL,
  colour_lim = NULL,
  colour_lim_contrast = NULL,
  show_single_contrast = FALSE,
  ...
)

\method{visualise}{ARA_discovery}(
  discovery,
  contrast = 1,
  metric = c("diff", "lfc"),
  mode = c("obsexp", "signal"),
  raw = FALSE,
  title = NULL,
  colour_lim = NULL,
  colour_lim_contrast = NULL,
  show_single_contrast = FALSE,
  ...
)

\method{visualise}{CS_discovery}(
  discovery,
  contrast = NULL,
  chr = "chr1",
  start = NULL,
  end = NULL,
  raw = FALSE,
  show_single_contrast = FALSE,
  ...
)

\method{visualise}{IS_discovery}(
  discovery,
  contrast = NULL,
  chr = "chr1",
  start = NULL,
  end = NULL,
  raw = FALSE,
  show_single_contrast = FALSE,
  ...
)

\method{visualise}{DI_discovery}(
  discovery,
  contrast = NULL,
  chr = "chr1",
  start = NULL,
  end = NULL,
  raw = FALSE,
  show_single_contrast = FALSE,
  ...
)

\method{visualise}{RCP_discovery}(
  discovery,
  contrast = 1,
  metric = c("smooth", "both", "lfc"),
  raw = F,
  title = NULL,
  flipFacet = F,
  ...
)

\method{visualise}{virtual4C_discovery}(
  discovery,
  bins = NULL,
  bedlist = NULL,
  bed_colours = "black",
  extend_viewpoint = NULL,
  ...
)

\method{visualise}{saddle_discovery}(
  discovery,
  contrast = 1,
  chr = "all",
  raw = FALSE,
  title = NULL,
  colour_lim = NULL,
  colour_lim_contrast = NULL,
  show_single_contrast = FALSE,
  ...
)

\method{visualise}{domainogram_discovery}(discovery, colour_lim = c(-1, 1), title = NULL, raw = FALSE, ...)

\method{visualise}{IIT_discovery}(
  discovery,
  contrast = 1,
  raw = FALSE,
  geom = c("boxplot", "violin", "jitter"),
  censor_contrast = TRUE,
  title = NULL,
  show_single_contrast = FALSE,
  ...
)
}
\arguments{
\item{discovery}{A \code{discovery} object as returned by GENOVA analysis
functions.}

\item{...}{further arguments passed to or from other methods.}

\item{contrast}{An \code{integer} or \code{character} matching an experiment
(name) of length 1, specifying which sample should be used to make a
contrast with all other samples. Alternatively, set to \code{NULL} to not
plot contrast panels. See also the \code{show_single_contrast} argument.}

\item{metric}{A \code{character} of length 1:

  \describe{ \item{A*A}{\code{"diff"} for difference by subtraction or
  \code{"lfc"} for \ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}} fold
  changes.} \item{RCP}{\code{"smooth"} for a
  \ifelse{html}{\out{log<sub>10</sub>}}{\eqn{log_10}}-smoothed line,
  \code{both} for adding the raw distance-bins as points and \code{lfc} for
  \ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}} fold changes.} }}

\item{raw}{A \code{logical} of length 1: should a bare bones plot be
returned?}

\item{title}{add a title}

\item{colour_lim, colour_lim_contrast}{One of: \itemize{ \item \code{NULL} to
have an educated quess for the scale. \item A \code{numeric} vector of
length two providing the limits of the scale. Use \code{NA} to refer to
existing minima or maxima. \item A \code{function} that accepts the
existing (automatic) limits and returns new limits. }}

\item{show_single_contrast}{A \code{logical} of length 1; if \code{FALSE}
(default), does not show contrasts when \code{discovery} describes one
experiment. If \code{TRUE}, plots empty panel.}

\item{mode}{\code{[PESCAn & ARA]} What result slot should be used for
visualisation? \code{"obsexp"} for the observed over expected metric or
\code{"signal"} for mean contacts at unshifted anchors.}

\item{chr}{\code{[CS, saddle & IS & DI]} A \code{character} of length 1
indicating a chromosome name.}

\item{start, end}{\code{[CS & IS & DI]} A \code{numeric} of length 1 with
start- and end-positions for the region to plot. If \code{NULL}, is set to
\code{-Inf} and \code{Inf} respectively.}

\item{flipFacet}{\code{[RCP]} Do you want to have RCP's of different regions
in one plot, instead of facets? (default : \code{FALSE})}

\item{bins}{\code{[virtual4C]} Set the number of histogram-bins}

\item{bedlist}{\code{[virtual4C]} Either a BED-formatted \code{data.frame} or
list thereof to plot underneath the tracks.}

\item{bed_colours}{\code{[virtual4C]} A \code{character} vector with colours
parallel to the number of list-elements in the \code{bedlist} argument.
Last colour is repeated if \code{bedlist} is longer than
\code{bed_colours}.}

\item{extend_viewpoint}{\code{[virtual4C]} Add a set bp to both sides of the
viewpoint. Makes the viewpoint-box broader.}

\item{geom}{\code{[IIT]} A \code{character} vector of length 1; either one of
\code{"boxplot"}, \code{"violin"}, \code{"jitter"} to get boxplots, violin
plots or jittered point plots.}

\item{censor_contrast}{\code{[IIT]} A \code{logical} of length 1 deciding
wether the contrasting experiment itself should be censored (\code{TRUE})
or included (\code{FALSE}).}
}
\description{
Plot the results of \code{discovery} objects. By default
  contrasts one sample with the others.
}
\details{
The \code{"diff"} \code{metric} value creates contrast panels by
  subtracting the values of each sample by the values of the sample indicated
  by the '\code{contrast}' argument. The \code{"lfc"} \code{metric} value
  creates contrast panels by dividing every samples' values by the sample
  indicated by the '\code{contrast}' argument and then taking the base 2
  logarithm of that division.

  The '\code{raw = TRUE}' argument allows custimisation of the plots. The
  returned plot will not have position- or fill-scales and no theme settings,
  which can be set freely afterwards to match personal aesthetic tastes. When
  '\code{raw = TRUE}' and '\code{subtract}' is not \code{NULL}, the fill
  scale of the contrast panels can be manipulated by setting the
  '\code{aesthetics = "altfill"}' inside ggplot2's fill scale functions.
}
\note{
For \code{ATA_discovery} objects which include the matrix's diagonal,
  the upper limit for the contacts fill scale is set to the 95th percentile
  of the data to increase the dynamic range of colours. The same is true for
  \code{ARA_discovery}, when '\code{mode = "signal"}'.
}
\examples{
\dontrun{
# APA
apa <- APA(list(WT = WT_40kb, KO = KO_40kb), loops)
visualise(apa)

# PE-SCAn
pescan <- PESCAn(list(WT = WT_40kb, KO = KO_40kb), super_enhancers)
visualise(pescan)

# To plot PE-SCAn without background correction
visualise(pescan, mode = "signal")

# ATA
ata <- APA(list(WT = WT_10kb, KO = KO_10kb), tads)
visualise(ata)

# ARA
ara <- ARA(list(WT = WT_20kb, KO = KO_20kb), ctcf_sites)
visualise(ara)

# Compartment score
cs <- compartment_score(list(WT = WT_100kb, KO = KO_100kb), H3K4me1_peaks)
visualise(cs, chr = "chr1")

# Saddle function
sadl <- saddle(list(WT = WT_100kb, KO = KO_100kb), cs) # see example above
visualise(sadl)

# Handling 'raw' plots
visualise(pescan, raw = TRUE) +
  ggplot2::scale_fill_gradient(aesthetics = "altfill")
}
}
