% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods_quantify.R,
%   R/methods_visualise.R
\name{visualise}
\alias{visualise}
\alias{quantify.APA_discovery}
\alias{visualise.default}
\alias{visualise.APA_discovery}
\alias{visualise.PESCAn_discovery}
\alias{visualise.ATA_discovery}
\alias{visualise.ARA_discovery}
\alias{visualise.RCP_discovery}
\alias{visualise.CS_discovery}
\title{Visualise discoveries}
\usage{
visualise(discovery, ...)

\method{quantify}{APA_discovery}(discovery, signal_size = 3)

\method{visualise}{APA_discovery}(discovery, contrast = 1,
  metric = c("diff", "lfc"), raw = FALSE, title = NULL)

\method{visualise}{PESCAn_discovery}(discovery, contrast = 1,
  metric = c("diff", "lfc"), mode = c("obsexp", "signal"),
  raw = FALSE, title = NULL)

\method{visualise}{ATA_discovery}(discovery, contrast = 1,
  metric = c("diff", "lfc"), raw = FALSE, title = NULL)

\method{visualise}{ARA_discovery}(discovery, contrast = 1,
  metric = c("diff", "lfc"), mode = c("obsexp", "signal"),
  raw = FALSE, title = NULL)

\method{visualise}{RCP_discovery}(discovery, contrast = 1,
<<<<<<< HEAD
  metric = c("smooth", "both", "lfc"), raw = F, flipFacet = F)

\method{visualise}{CS_discovery}(discovery, contrast = NULL,
  chr = "chr1", start = NULL, end = NULL, raw = FALSE)
=======
  metric = c("smooth", "both", "lfc"), raw = F, title = NULL,
  flipFacet = F)
>>>>>>> 3748225d5a6e213f603033d0c56a180b41f7955a
}
\arguments{
\item{discovery}{A \code{discovery} object as returned by GENOVA analysis
functions.}

\item{contrast}{An \code{integer} or \code{character} matching an experiment
(name) of length 1, specifying which sample should be used to make a
contrast with all other samples. Alternatively, set to \code{NULL} to not
plot contrast panels.}

\item{metric}{A \code{character} of length 1:

  \describe{ \item{A*A}{\code{"diff"} for difference by subtraction or
  \code{"lfc"} for \ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}} fold
  changes.} \item{RCP}{\code{"smooth"} for a log10-smoothed line, \code{both}
  for adding the raw distance-bins as points and \code{lfc} for
  \ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}} fold changes.} }}

\item{raw}{A \code{logical} of length 1: should a bare bones plot be
returned?}

\item{mode}{(PESCAn_discovery and ARA_discovery only) What result slot should
be used for visualisation? \code{"obsexp"} for the observed over expected
metric or \code{"signal"} for mean contacts at unshifted anchors.}

\item{flipFacet}{(RCP_discovery only) Do you want to have RCP's of different
regions in one plot, instead of facets? (default : \code{FALSE})}

\item{chr}{(CP_discovery only) A \code{character} of length 1 indicating a
chromosome name.}

\item{start, end}{(CP_discovery only) A \code{numeric} of length 1 with start-
and end-positions for the region to plot. If \code{NULL}, is set to
\code{-Inf} and \code{Inf} respectively.}
}
\description{
Plot the results of \code{discovery} objects. By default
  contrasts one sample with the others.
}
\details{
The \code{"diff"} \code{metric} value creates contrast panels by
  subtracting the values of each sample by the values of the sample indicated
  by the '\code{contrast}' argument. The \code{"lfc"} \code{metric} value
  creates contrast panels by dividing every samples' values by the sample
  indicated by the '\code{contrast}' argument and then taking the base 2
  logarithm of that division.

  The '\code{raw = TRUE}' argument allows custimisation of the plots. The
  returned plot will not have position- or fill-scales and no theme settings,
  which can be set freely afterwards to match personal aesthetic tastes. When
  '\code{raw = TRUE}' and '\code{subtract}' is not \code{NULL}, the fill
  scale of the contrast panels can be manipulated by setting the
  '\code{aesthetics = "altfill"}' inside ggplot2's fill scale functions.
}
\note{
For \code{ATA_discovery} objects which include the matrix's diagonal,
  the upper limit for the contacts fill scale is set to the 95th percentile
  of the data to increase the dynamic range of colours. The same is true for
  \code{ARA_discovery}, when '\code{mode = "signal"}'.
}
\examples{
# APA
apa <- APA(list(WT = WT_40kb, KO = KO_40kb), loops)
visualise(apa)

# PE-SCAn
pescan <- PESCAn(list(WT = WT_40kb, KO = KO_40kb), super_enhancers)
visualise(pescan)

# To plot PE-SCAn without background correction
visualise(pescan, mode = "signal")

# ATA
ata <- APA(list(WT = WT_10kb, KO = KO_10kb), tads)
visualise(ata)

# ARA
ara <- ARA(list(WT = WT_20kb, KO = KO_20kb), ctcf_sites)
visualise(ara)

# Handling 'raw' plots
visualise(pescan, raw = TRUE) +
  ggplot2::scale_fill_gradient(aesthetics = "altfill")
}
