% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_visualise.R
\name{visualise}
\alias{visualise}
\alias{visualize}
\alias{visualise.default}
\alias{visualise.APA_discovery}
\alias{visualise.CSCAn_discovery}
\alias{visualise.PESCAn_discovery}
\alias{visualise.ATA_discovery}
\alias{visualise.ARA_discovery}
\alias{visualise.CS_discovery}
\alias{visualise.IS_discovery}
\alias{visualise.DI_discovery}
\alias{visualise.RCP_discovery}
\alias{visualise.virtual4C_discovery}
\alias{visualise.saddle_discovery}
\alias{visualise.domainogram_discovery}
\alias{visualise.IIT_discovery}
\alias{visualise.chrommat_discovery}
\title{Visualise discoveries}
\usage{
visualise(discovery, ...)

\method{visualise}{default}(
  discovery,
  contrast,
  raw,
  title,
  colour_lim,
  colour_lim_contrast,
  ...
)
}
\arguments{
\item{discovery}{A \code{discovery} object as returned by GENOVA analysis
functions.}

\item{...}{Further arguments specific to the discovery class. See the section 
extended arguments below.}

\item{contrast}{An \code{integer} or \code{character} matching an experiment
(name) of length 1, specifying which sample should be used to make a
contrast with all other samples. Alternatively, set to \code{NULL} to not
plot contrast panels. See also the \code{show_single_contrast} argument.}

\item{raw}{A \code{logical} of length 1: should a bare bones plot be
returned?}

\item{title}{add a title}

\item{colour_lim, colour_lim_contrast}{Indication of limits for the primary and secondary continuous colour scales 
respectively. One of: \itemize{ 
\item \code{NULL} to have an educated quess for the scale. 
\item A \code{numeric} vector of length two providing the limits of the 
scale. Use \code{NA} to refer to existing minima or maxima. 
\item A \code{function} that accepts the existing (automatic) limits and 
returns new limits.}}
}
\description{
Plot the results of \code{discovery} objects. By default
  contrasts one sample with the others.
}
\details{
The \code{"diff"} \code{metric} value creates contrast panels by
  subtracting the values of each sample by the values of the sample indicated
  by the '\code{contrast}' argument. The \code{"lfc"} \code{metric} value
  creates contrast panels by dividing every samples' values by the sample
  indicated by the '\code{contrast}' argument and then taking the base 2
  logarithm of that division.

  The '\code{raw = TRUE}' argument allows custimisation of the plots. The
  returned plot will not have position- or fill-scales and no theme settings,
  which can be set freely afterwards to match personal aesthetic tastes. When
  '\code{raw = TRUE}' and '\code{subtract}' is not \code{NULL}, the fill
  scale of the contrast panels can be manipulated by setting the
  '\code{aesthetics = "altfill"}' inside ggplot2's fill scale functions.
}
\note{
For \code{ATA_discovery} objects which include the matrix's diagonal,
  the upper limit for the contacts fill scale is set to the 95th percentile
  of the data to increase the dynamic range of colours. The same is true for
  \code{ARA_discovery}, when '\code{mode = "signal"}'.
}
\section{Extended arguments}{


\subsection{APA, PE-SCAn, ATA, ARA & C-SCAn}{
\describe{
 \item{\code{metric}}{A \code{character} of length one: \code{"diff"} for 
  difference by subtraction or \code{"lfc"} for 
  \ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}} fold changes.}
 \item{\code{colour_lim}, \code{colour_lim_contrast}}{
  Indication of limits for the primary and secondary colour scale 
  respectively. One of: \itemize{ 
  \item \code{NULL} to have an educated quess for the scale. 
  \item A \code{numeric} vector of length two providing the limits of the 
  scale. Use \code{NA} to refer to existing minima or maxima. 
  \item A \code{function} that accepts the existing (automatic) limits and 
  returns new limits.}}
 \item{\code{mode}}{A \code{character} of length one indicating what type of
 result to plot. Either \code{"signal"} or \code{"obsexp"}, referring to the
 slot in the discovery object. Applicatble to PE-SCAn, C-SCAn and ARA.}
 \item{\code{show_single_contrast}}{A \code{logical} of length 1; 
  if \code{FALSE} (default), does not show contrasts when \code{discovery} 
  describes one experiment. If \code{TRUE}, plots empty panel.}
}
}

\subsection{Compartment-, Insulation score & 
Directionality Index}{
\describe{
 \item{\code{chr}}{A \code{character} of length one with the chromosome 
 name. Defaults to \code{"chr1"}.}
 \item{\code{start}, \code{end}}{An \code{integer} of length one setting the 
 start or end of the region to plot. If \code{NULL} (default), is set to 
 \code{-Inf} and \code{Inf} respectively}
}
}

\subsection{Relative Contact Probability}{
\describe{
 \item{\code{metric}}{A \code{character} of length one. The choices are:
 \describe{
  \item{\code{"smooth"}}{A 
  \ifelse{html}{\out{log<sub>10</sub>}}{\eqn{log_10}}-smoothed line 
  (default)}
  \item{\code{"both"}}{Like \code{"smooth"}, but also adds the raw binned 
  distances as points.}
  \item{\code{"lfc"}}{Displays 
  \ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}} fold change compared to 
  the sample specified in the \code{contrast} argument.}
 }}
 \item{\code{flipFacet}}{A \code{logical} of length one. If the 
 \code{bedlist} argument was provided to \code{RCP()}, combine all regions 
  in one panel? Defaults to \code{FALSE}.}
}
}

 \subsection{Virtual 4C}{
\describe{
 \item{\code{bins}}{An \code{integer} of length 1 with the number of bins to
  aggregate signal in. If \code{NULL} (the default), this is set to the 
  number of Hi-C bins in the viewpoint's chromosome.}
 \item{\code{bedlist}}{Either a BED-formatted \code{data.frame} or a 
 \code{list} thereof, indicating genomic intervals to annotate in the bottom 
 margin of the plot.}
 \item{\code{extend_viewpoint}}{An \code{integer} of length one in basepairs
 indicating by how much to widen the viewpoint censor-box. Affects the 
 scaling of the y-axis.}
}
}

\subsection{Saddle}{
\describe{
 \item{\code{chr}}{A \code{character} of length one with the chromosome 
  name. Defaults to \code{"all"}.}
 \item{\code{show_single_contrast}}{A \code{logical} of length 1; 
  if \code{FALSE} (default), does not show contrasts when \code{discovery} 
  describes one experiment. If \code{TRUE}, plots empty panel.}
}
}

\subsection{Intra-inter TAD}{
\describe{
 \item{\code{geom}}{One of the following length one \code{character}s:
   \describe{
     \item{\code{"boxplot"}}{Display as boxplots.}
     \item{\code{"violin"}}{Display as violin plots.}
     \item{\code{"jitter"}}{Display as jittered points plot.}
   }
 }
 \item{\code{censor_contrast}}{A \code{logical} of length 1 deciding whether 
 the contrasting experiment itself should be censored (\code{TRUE}) or
 included (\code{FALSE}).}
}
}
}

\examples{
\dontrun{
# APA
apa <- APA(list(WT = WT_40kb, KO = KO_40kb), loops)
visualise(apa)

# PE-SCAn
pescan <- PESCAn(list(WT = WT_40kb, KO = KO_40kb), super_enhancers)
visualise(pescan)

# To plot PE-SCAn without background correction
visualise(pescan, mode = "signal")

# ATA
ata <- APA(list(WT = WT_10kb, KO = KO_10kb), tads)
visualise(ata)

# ARA
ara <- ARA(list(WT = WT_20kb, KO = KO_20kb), ctcf_sites)
visualise(ara)

# Compartment score
cs <- compartment_score(list(WT = WT_100kb, KO = KO_100kb), H3K4me1_peaks)
visualise(cs, chr = "chr1")

# Saddle function
sadl <- saddle(list(WT = WT_100kb, KO = KO_100kb), cs) # see example above
visualise(sadl)

# Handling 'raw' plots
visualise(pescan, raw = TRUE) +
  ggplot2::scale_fill_gradient(aesthetics = "altfill")
}
}
