% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddle.R
\name{visualise.saddle}
\alias{visualise.saddle}
\title{Plot saddle-plots}
\usage{
visualise.saddle(SBoutList, addText = T, zlim = c(-0.5, 0.5),
  EVlim = NULL, square = T, crossLines = NULL)
}
\arguments{
\item{SBoutList}{A list of outputs trom saddleBins.}

\item{addText}{A list of outputs trom saddleBins.}

\item{zlim}{Set the z-lims of the matrix.}

\item{EVlim}{If set, use a different set of z-lims for the compartment-scores.}

\item{square}{Set this to FALSE if plots are unaligned: will not produce a rigidly square matrix.}

\item{crossLines}{Plot crosslines in matrix. Crossline denote border of- A and B-bins. If not set, will plot them when nBins  >= 10.}
}
\value{
A saddle-plot
}
\description{
Takes the input from saddleBins and produces saddleplots.
}
