% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddleFunctions.R
\name{visualise.saddle}
\alias{visualise.saddle}
\title{Plot saddle-plots}
\usage{
visualise.saddle(SBoutList, addText = T, zlim = c(0.5, 2),
  EVlim = c(-1.5, 1.5), square = T, crossLines = NULL)
}
\arguments{
\item{SBoutList}{A list of outputs trom saddle().}

\item{addText}{Add annotations in the corners for legibility.}

\item{zlim}{Set the z-lims of the matrix.}

\item{EVlim}{Set the z-lims for the average compartment-scores on top.}

\item{square}{Set this to FALSE if plots are unaligned: will not produce a rigidly square matrix.}

\item{crossLines}{Plot crosslines in matrix. Crossline denote border of- A and B-bins. If not set, will plot them when nBins  >= 10.}
}
\value{
A saddle-plot
}
\description{
Takes the input from saddle() and produces saddleplots.
Also plot a lowess-smoothed summary track of the sorted compartment-scores on top.
}
\examples{
# run saddle() on all chromosomes with 25 bins.
saddle_WT = saddle(exp = Hap1_WT_1MB, chip = H3K27acPeaks, nBins = 25)

# plot saddle-plot
visualise.saddle(list(saddle_WT), crossLines = T, addText = T)
}
